% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataset.R
\name{loadCountMatrix}
\alias{loadCountMatrix}
\title{loadCountMatrix}
\usage{
loadCountMatrix(file, header, sep, dec)
}
\arguments{
\item{file}{Path of the count matrix to load.}

\item{header}{Set TRUE if the first row of the table corresponds to the
column names, and FALSE if it doesn't.}

\item{sep}{Character used in the table to separate the fields. 
Usually it's ' ' ';' ',' or '\\t'.}

\item{dec}{Character used in the table for decimal points.
Usually '.' or ',' .}
}
\value{
The count matrix.
}
\description{
This function allows to import the count matrix. All parameters are 
mandatory and no default values, so it is necessary to know how is your 
count matrix before importing.
}
\examples{

countmatrixPath <- system.file("extdata/Bergiers_counts_matrix_filtered.tsv",
                                package = "conclus")

loadCountMatrix(file=countmatrixPath, header=TRUE, sep='\t', dec=".")

}
