% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-clustering.R
\name{addClustering,scRNAseq-method}
\alias{addClustering,scRNAseq-method}
\alias{addClustering}
\title{addClustering}
\usage{
addClustering(theObject, filePathAdd=NA, clusToAdd=NA)
}
\arguments{
\item{theObject}{An Object of class scRNASeq for which
?calculateClustersSimilarity was used.}

\item{filePathAdd}{Default=NA. Path to the file containing the clustering to
replace in the object. It should be made of two columns 'clusters' and
'cells'. This should be left to NA if clusToAdd is used. Default=NA.}

\item{clusToAdd}{Data frame having two columns 'clusters' and 'cells'
containing the clustering to be used in theObject. Should be left to NA if
filePathAdd is used. Default=NA.}
}
\value{
An object of class scRNASeq with its column name metadata updated.
}
\description{
This method enables to add a clustering to the existing object in order to
change the coloration of the t-sne. It is particularly useful to compare the
performance of different tools.
}
\examples{
## Object scr containing the results of previous steps
load(system.file("extdata/example_scrFull.Rdat", package="conclus"))

## Retrieving the table indicating to which cluster each cell belongs
clustCellsDf <- retrieveTableClustersCells(scr)

## Replace “4” by “3” to merge 3/4
clustCellsDf$clusters[which(clustCellsDf$clusters == 4)] <- 3

## Modifying the object to take into account the new classification
scrUpdated <- addClustering(scr, clusToAdd=clustCellsDf)


}
\author{
Ilyess RACHEDI, based on code by Polina PAVLOVICH and Nicolas DESCOSTES.
}
