% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_and_clustering_functions.R
\name{addClusteringManually}
\alias{addClusteringManually}
\title{addClusteringManually}
\usage{
addClusteringManually(fileName, sceObject, dataDirectory, experimentName,
  columnName = "clusters")
}
\arguments{
\item{fileName}{a file with the clustering solution (for example, from previous CONCLUS runs).}

\item{sceObject}{a SingleCellExperiment object with your experiment.}

\item{dataDirectory}{output directory (supposed to be the same for one experiment during the workflow).}

\item{experimentName}{name of the experiment which appears in filenames (supposed to be the same for one experiment during the workflow).}

\item{columnName}{name of the column with the clusters.}
}
\value{
A SingleCellExperiment object with the created/renewed column "clusters" in the colData(sceObject).
}
\description{
The function replaces the content of the column "clusters" in the colData(sceObject) 
with the clustering provided in the user table.
The function will return the sceObject with cells which intersect with the cells from the input table.
}
