% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{assignClusters}
\alias{assignClusters}
\title{From conclus object, assign cluster memberships to all items in a dissimilarity matrix}
\usage{
assignClusters(d, conclus, k)
}
\arguments{
\item{d}{A dissimilarity matrix containing a full dataset from which a subset
has been used in \code{conclus}.}

\item{conclus}{An object of class 'conclus'.}

\item{k}{The number of clusters you believe there to be.}
}
\description{
From conclus object, assign cluster memberships to all items in a dissimilarity matrix
}
\details{
With 'large' datasets, dissimilarity matrices get big and it is
  sometimes useful to do clustering on the basis of a subset, to conserve
  memory. Once clusters have been identified, the cluster members of the items
  not in the subset might be required. This function obtains them by finding
  the representative items, on the basis of item consensus, and allocating
  other items on the basis of their minum distance from the represenetatives.
}
\seealso{
\code{\link{consensus}}, \code{\link{representatives}}
}
