% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-clustering.R
\name{calculateClustersSimilarity,scRNAseq-method}
\alias{calculateClustersSimilarity,scRNAseq-method}
\alias{calculateClustersSimilarity}
\title{calculateClustersSimilarity}
\usage{
calculateClustersSimilarity(theObject, clusteringMethod = "ward.D2")
}
\arguments{
\item{theObject}{An Object of class scRNASeq for which the count matrix was
normalized (see ?normaliseCountMatrix), tSNE were calculated (see
?generateTSNECoordinates), dbScan was run (see ?runDBSCAN), and cells were
clustered (see ?clusterCellsInternal).}

\item{clusteringMethod}{Clustering method passed to hclust() function. See
?hclust for a list of method. Default = "ward.D2".}
}
\value{
An object of class scRNASeq with its clustersSimilarityMatrix and
clustersSimiliratyOrdered slots updated.
}
\description{
Having computed cells similarity, pools information into clusters.
}
\examples{
## Object scr containing the results of previous steps
load(system.file("extdata/example_scrFull.Rdat", package="conclus"))
## Calculate clusters similarity
scr <- calculateClustersSimilarity(scr)

}
\seealso{
plotClustersSimilarity
}
\author{
Ilyess RACHEDI, based on code by Polina PAVLOVICH and Nicolas DESCOSTES.
}
