% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_and_clustering_functions.R
\name{calculateClustersSimilarity}
\alias{calculateClustersSimilarity}
\title{Having cells similarity, calculate clusters similarity.}
\usage{
calculateClustersSimilarity(cellsSimilarityMatrix, sceObject,
  clusteringMethod)
}
\arguments{
\item{cellsSimilarityMatrix}{a similarity matrix, one of the results of conclus::clusterCellsInternal() function.}

\item{sceObject}{a SingleCellExperiment object with your experiment.}

\item{clusteringMethod}{a clustering methods passed to hclust() function.}
}
\value{
A list contating the cluster similarity matrix and cluster names (order).
}
\description{
Having cells similarity, calculate clusters similarity.
}
