% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_and_clustering_functions.R
\name{choosePalette}
\alias{choosePalette}
\title{Choose palette for a plot.}
\usage{
choosePalette(colorPalette, clustersNumber)
}
\arguments{
\item{colorPalette}{Either "default" or a vector of colors, for example c("yellow", "#CC79A7").}

\item{clustersNumber}{number of clusters in the output palette.}
}
\value{
Color palette with the number of colors equal to the clusterNumber parameter.
}
\description{
It is an internal function usually applied for choosing the palette for clusters.
Depending if the number of clusters is more than 12 or not, one of two built-in palettes will be applied.
If you give your vector of colors, the function will not change them.
If the number of clusters is more than 26, it will copy colors to get the needed length of the palette.
}
