% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_and_clustering_functions.R
\name{clusterCellsInternal}
\alias{clusterCellsInternal}
\title{Cluster cells and get similarity matrix of cells.}
\usage{
clusterCellsInternal(dbscanMatrix, sceObject, clusterNumber = 0,
  deepSplit, cores = 14, clusteringMethod = "ward.D2")
}
\arguments{
\item{dbscanMatrix}{an output matrix of conclus::runDBSCAN() function.}

\item{sceObject}{a SingleCellExperiment object with your experiment.}

\item{clusterNumber}{a parameter, specifying the exact number of cluster.}

\item{deepSplit}{a parameter, specifying how deep we will split the clustering tree. It takes integers from 1 to 4.}

\item{cores}{maximum number of jobs that CONCLUS can run in parallel.}

\item{clusteringMethod}{a clustering methods passed to hclust() function.}
}
\value{
A SingleCellExperiment object with modified/created "clusters" column in the colData, and cells similarity matrix.
}
\description{
The function returns consensus clusters by using hierarchical clustering on the similarity matrix of cells.
It provides two options: to specify an exact number of clusters (with clusterNumber parameter)
or to select the depth of splitting (deepSplit parameter).
}
