% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFormatting.R
\name{conclusCacheClear}
\alias{conclusCacheClear}
\title{conclusCacheClear}
\usage{
conclusCacheClear()
}
\value{
Nothing, it deletes the cache of conclus.
}
\description{
This function deletes the cache of conclus.
}
\details{
This function don't return anything. It deletes the current contents
of the cache.
}
\examples{
cache <- tools::R_user_dir("conclus", which="cache")
bfc <- BiocFileCache::BiocFileCache(cache, ask = FALSE)

conclusCacheClear()

}
\author{
Ilyess RACHEDI & Nicolas DESCOSTES
}
