% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{consensus}
\alias{consensus}
\title{Get cluster and item consensus indices}
\usage{
consensus(x, type = "cluster")
}
\arguments{
\item{x}{An object of class 'consclus'}

\item{type}{Character string taking values 'cluster' or 'item'}
}
\description{
Get cluster and item consensus indices
}
\details{
If \code{type='cluster'} a matrix with the cluster consensus values
  is returned. A \code{print} method exists. If \code{type='item'} an array
  is returned with as many rows as items. To access the item consensus for,
  say, k=4, use \code{x[,,3]} (remembering that the array extent starts with
  k=2). The item consensus values give a measure of belongingness to each item
  for each cluster. Prototypes can be extracted by finding the item with the
  largest consensus for each cluster.
}
\seealso{
\code{\link{representatives}}, \code{\link{conclus}}, \code{\link{assignClusters}}
}
