% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{constructors}
\alias{constructors}
\alias{singlecellRNAseq}
\title{constructors}
\usage{
singlecellRNAseq(experimentName, countMatrix, species, outputDirectory)
}
\arguments{
\item{experimentName}{character string representing the name of the
experiment. Many output of scRNAseq will use this name.}

\item{countMatrix}{An integer matrix representing the raw count matrix with
reads or unique molecular identifiers (UMIs).}

\item{species}{character string representing the species of interest.
Currently limited to "mouse" and "human".}

\item{outputDirectory}{A character string of the path to the root output
folder.}
}
\value{
Object of class scRNAseq
}
\description{
Constructor of the class scRNAseq (might be extended in the future).
}
\examples{
experimentName <- "Bergiers"
countMatrix <- as.matrix(read.delim(system.file(
"extdata/test_countMatrix.tsv", package="conclus")))
outputDirectory <- "YourOutputDirectory"
columnsMetaData <- read.delim(
system.file("extdata/Bergiers_colData_filtered.tsv", package="conclus"))

## Create the initial object
scr <- singlecellRNAseq(experimentName = experimentName,
                countMatrix     = countMatrix,
                species         = "mouse",
                outputDirectory = outputDirectory)
                
}
\seealso{
scRNAseq-class
}
