% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Dbscan-class}
\alias{Dbscan-class}
\alias{Dbscan}
\title{The Dbscan class}
\description{
S4 class. This class represents the features of a DBSCAN
clustering, like the minPoints and the espilon values used to make the
clusters, and the results of the clustering.
}
\section{Slots}{

 \describe{
   \item{\code{name}:}{Variable of class \code{"character"},
   It's the name of the DBSCAN clustering.}
   \item{\code{epsilon}:}{Vector of class \code{"numeric"}. The espsilon is
   the distance to consider two points belonging to the same cluster.
   By default, CONCLUS use a range of epsilon= c(1.3, 1.4, 1.5), so the value
   is one of them.}
   \item{\code{minPoints}:}{Vector of class \code{"numeric"}. The minPoints
   is the minimum number of points to construct a cluster. By default,
   CONCLUS use a range of minPoints= c(3, 4), so the value
   is one of them.}
   \item{\code{clustering}:}{Variable of class \code{"matrix"}, it contains
   the result of one DBSCAN clustering solution.}
   }
}

