% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_and_clustering_functions.R
\name{exportClusteringResults}
\alias{exportClusteringResults}
\title{exportClusteringResults}
\usage{
exportClusteringResults(sceObject, dataDirectory, experimentName, fileName)
}
\arguments{
\item{sceObject}{a SingleCellExperiment object with your experiment.}

\item{dataDirectory}{output directory (supposed to be the same for one experiment during the workflow).}

\item{experimentName}{name of the experiment which appears at the beginning of the file name 
(supposed to be the same for one experiment during the workflow).}

\item{fileName}{the rest of output file name.}
}
\description{
The function saves clustering results into a table. Row names are cell names in the same order as in the sceObject.
}
