% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_and_clustering_functions.R
\name{exportMatrix}
\alias{exportMatrix}
\title{Export matrix to a file.}
\usage{
exportMatrix(matrix, dataDirectory, experimentName, name)
}
\arguments{
\item{matrix}{your matrix (e.g., expression matrix)}

\item{dataDirectory}{CONCLUS output directory for a given experiment (supposed to be the same for one experiment during the workflow).}

\item{experimentName}{name of the experiment which will appear at the beginning of the filenames 
(supposed to be the same for one experiment during the workflow).}

\item{name}{name of the file. Will be placed after the experimentName header.}
}
\description{
The function allows you to export a matrix to a .csv file with a hard-coded filename (according to experimentName) 
in the "dataDirectory/output_tables" directory for further analysis.
}
