% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conclus-package.R
\docType{data}
\name{Gaussian3}
\alias{Gaussian3}
\alias{Uniform1}
\title{Gaussian3 and Uniform1 simulated data}
\format{Data frames of simulated data, described in Monti et al.}
\source{
http://portals.broadinstitute.org/cgi-bin/cancer/publications/view/87 Beware! the URL
  has been known to change. An alternative is to go to http://portals.broadinstitute.org/cgi-bin/cancer
  and search for the paper by Monti et al.
}
\description{
Gaussian3 and Uniform1 simulated data
}
\references{
S. Monti, P. Tamayo, J. Mesirov and T. Golub, Consensus clustering: a
  resampling-based method for class discovery and visualization of gene expression data,
  Machine Learning, 52, 91 -- 118, 2003
}
