% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_and_clustering_functions.R
\name{generateTSNECoordinates}
\alias{generateTSNECoordinates}
\title{Generate and save t-SNE coordinates with selected parameters.}
\usage{
generateTSNECoordinates(sceObject, dataDirectory, experimentName,
  randomSeed = 42, cores = 14, PCs = c(4, 6, 8, 10, 20, 40, 50),
  perplexities = c(30, 40))
}
\arguments{
\item{sceObject}{a SingleCellExperiment object with your experiment.}

\item{dataDirectory}{output directory for CONCLUS (supposed to be the same for one experiment during the workflow).}

\item{experimentName}{name of the experiment which will appear in filenames (supposed to be the same for one experiment during the workflow).}

\item{randomSeed}{random seed for reproducibility.}

\item{cores}{maximum number of jobs that CONCLUS can run in parallel.}

\item{PCs}{a vector of first principal components.
For example, to take ranges 1:5 and 1:10 write c(5, 10).}

\item{perplexities}{a vector of perplexity (t-SNE parameter).}
}
\value{
An object with t-SNE results (coordinates for each plot).
}
\description{
The function generates several t-SNE coordinates based on given perplexity and ranges of PCs. 
Final number of t-SNE plots is length(PCs)*length(perplexities)
It writes coordinates in "dataDirectory/tsnes" subfolder.
}
