% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\name{getCellsSimilarityMatrix}
\alias{getCellsSimilarityMatrix}
\alias{getCellsSimilarityMatrix,scRNAseq-method}
\alias{getCellsSimilarityMatrix,getCellsSimilarityMatrix-method}
\title{Getter of the cellsSimilarityMatrix slot of scRNAseq class}
\usage{
getCellsSimilarityMatrix(theObject)

\S4method{getCellsSimilarityMatrix}{scRNAseq}(theObject)
}
\arguments{
\item{theObject}{scRNAseq object}
}
\value{
Cells Similarity Matrix
}
\description{
Get a Cells Similarity Matrix
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
