% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\name{getClustering}
\alias{getClustering}
\alias{getClustering,Dbscan-method}
\alias{getClustering,getClustering-method}
\title{Getter of the clustering slot of Dbscan class}
\usage{
getClustering(theObject)

\S4method{getClustering}{Dbscan}(theObject)
}
\arguments{
\item{theObject}{Dbscan object}
}
\value{
clustering result of the Dbscan object
}
\description{
Get the clustering from the Dbscan object
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
