% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\name{getClustersMarkers}
\alias{getClustersMarkers}
\alias{getClustersMarkers,scRNAseq-method}
\alias{getClustersMarkers,getClustersMarkers-method}
\title{Getter of the clustersMarkers slot of scRNAseq class}
\usage{
getClustersMarkers(theObject)

\S4method{getClustersMarkers}{scRNAseq}(theObject)
}
\arguments{
\item{theObject}{scRNAseq object}
}
\value{
Clusters markers
}
\description{
Get the clusters markers
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
