% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\name{getClustersSimilarityMatrix}
\alias{getClustersSimilarityMatrix}
\alias{getClustersSimilarityMatrix,scRNAseq-method}
\alias{getClustersSimilarityMatrix,getClustersSimilarityMatrix-method}
\title{Getter of the clusters similarity matrix slot of scRNAseq class}
\usage{
getClustersSimilarityMatrix(theObject)

\S4method{getClustersSimilarityMatrix}{scRNAseq}(theObject)

\S4method{getClustersSimilarityMatrix}{scRNAseq}(theObject)
}
\arguments{
\item{theObject}{scRNAseq object}
}
\value{
clusters similarity matrix
}
\description{
Get clusters similarity matrix
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
