% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getClustersSimiliratyOrdered}
\alias{getClustersSimiliratyOrdered}
\title{Getter of the clustersSimiliratyOrdered slot of scRNAseq class}
\usage{
getClustersSimiliratyOrdered(theObject)
}
\arguments{
\item{theObject}{scRNAseq object}
}
\value{
Factor. Order of the consensus clusters.
}
\description{
Get the order of the consensus clusters
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
