% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\name{getCoordinates}
\alias{getCoordinates}
\alias{getCoordinates,Tsne-method}
\alias{getCoordinates,getCoordinates-method}
\title{Getter of the coordinates slot of Tsne class}
\usage{
getCoordinates(theObject)

\S4method{getCoordinates}{Tsne}(theObject)
}
\arguments{
\item{theObject}{Tsne object}
}
\value{
coordinates of the Tsne object
}
\description{
Get the coordinates from the Tsne object
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
