% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\name{getCountMatrix}
\alias{getCountMatrix}
\alias{getCountMatrix,scRNAseq-method}
\alias{getCountMatrix,getCountMatrix-method}
\title{Getter of countMatrix slot of scRNAseq class}
\usage{
getCountMatrix(theObject)

\S4method{getCountMatrix}{scRNAseq}(theObject)
}
\arguments{
\item{theObject}{scRNAseq object}
}
\value{
The matrix assigned to the countMatrix slot.
}
\description{
Get the the count matrix assigned to the countMatrix slot.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
