% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\name{getDbscanList}
\alias{getDbscanList}
\alias{getDbscanList,scRNAseq-method}
\alias{getDbscanList,getDbscanList-method}
\title{Getter of the dbscanList slot of scRNAseq class}
\usage{
getDbscanList(theObject)

\S4method{getDbscanList}{scRNAseq}(theObject)
}
\arguments{
\item{theObject}{scRNAseq object}
}
\value{
List of DBSCAN results
}
\description{
Get a list of DBSCAN results
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
