% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\name{getDbscanName}
\alias{getDbscanName}
\alias{getDbscanName,Dbscan-method}
\alias{getDbscanName,getDbscanName-method}
\title{Getter of the name slot of Dbscan class}
\usage{
getDbscanName(theObject)

\S4method{getDbscanName}{Dbscan}(theObject)
}
\arguments{
\item{theObject}{Dbscan object}
}
\value{
Name of the Dbscan object
}
\description{
Get the name from the Dbscan object
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
