% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\name{getEpsilon}
\alias{getEpsilon}
\alias{getEpsilon,Dbscan-method}
\alias{getEpsilon,getEpsilon-method}
\title{Getter of the epsilon slot of Dbscan class}
\usage{
getEpsilon(theObject)

\S4method{getEpsilon}{Dbscan}(theObject)
}
\arguments{
\item{theObject}{Dbscan object}
}
\value{
Name of the Dbscan object
}
\description{
Get the name from the Dbscan object
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
