% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\name{getExperimentName}
\alias{getExperimentName}
\alias{getExperimentName,scRNAseq-method}
\alias{getExperimentName,getExperimentName-method}
\title{Getter of experimentName slot of scRNAseq class}
\usage{
getExperimentName(theObject)

\S4method{getExperimentName}{scRNAseq}(theObject)
}
\arguments{
\item{theObject}{scRNAseq object}
}
\value{
a string corresponding to the experiment name
}
\description{
Get the experiment name assigned to the experimentName slot. Experiment name
describes the name of th experiment of the Conclus analysis.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
