% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\name{getGenesInfos}
\alias{getGenesInfos}
\alias{getGenesInfos,scRNAseq-method}
\alias{getGenesInfos,getGenesInfos-method}
\title{Getter of the getGenesInfos slot of scRNAseq class}
\usage{
getGenesInfos(theObject)

\S4method{getGenesInfos}{scRNAseq}(theObject)
}
\arguments{
\item{theObject}{scRNAseq object}
}
\value{
Data.frame containing informations about markers genes.
}
\description{
Get the clusters markers from the object
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
