% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_and_clustering_functions.R
\name{getMarkerGenes}
\alias{getMarkerGenes}
\title{Get top N marker genes from each cluster.}
\usage{
getMarkerGenes(dataDirectory, sceObject, genesNumber = 14,
  experimentName, removeDuplicates = TRUE)
}
\arguments{
\item{dataDirectory}{output directory for a run of CONCLUS (supposed to be the same for one experiment during the workflow).}

\item{sceObject}{a SingleCellExperiment object with your experiment.}

\item{genesNumber}{top N number of genes to get from one cluster.}

\item{experimentName}{name of the experiment which appears in filenames (supposed to be the same for one experiment during the workflow).}

\item{removeDuplicates}{boolean, if duplicated genes must be deleted or not.}
}
\value{
A data frame where the first columns are marker genes ("geneName") and 
the second column is the groups ("clusters").
}
\description{
This function reads results of conclus::rankGenes() from "dataDirectory/marker_genes" and selects top N markers for each cluster.
}
