% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\name{getMarkerGenesList}
\alias{getMarkerGenesList}
\alias{getMarkerGenesList,scRNAseq-method}
\alias{getMarkerGenesList,getExperimentName-method}
\title{Getter of the markerGenesList slot of scRNAseq class}
\usage{
getMarkerGenesList(theObject)

\S4method{getMarkerGenesList}{scRNAseq}(theObject)
}
\arguments{
\item{theObject}{scRNAseq object}
}
\value{
markerGenesList
}
\description{
Get the order of the consensus clusters
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
