% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\name{getMinPoints}
\alias{getMinPoints}
\alias{getMinPoints,Dbscan-method}
\alias{getMinPoints,getExperimentName-method}
\title{Getter of the minPoints slot of Dbscan class}
\usage{
getMinPoints(theObject)

\S4method{getMinPoints}{Dbscan}(theObject)
}
\arguments{
\item{theObject}{Dbscan object}
}
\value{
minPoints of the Dbscan object
}
\description{
Get the minPoints from the Dbscan object
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
