% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\name{getOutputDirectory}
\alias{getOutputDirectory}
\alias{getOutputDirectory,scRNAseq-method}
\alias{getOutputDirectory,getOutputDirectory-method}
\title{Getter of the outputDirectory slot of scRNAseq class}
\usage{
getOutputDirectory(theObject)

\S4method{getOutputDirectory}{scRNAseq}(theObject)
}
\arguments{
\item{theObject}{scRNAseq object}
}
\value{
The path of the output directory
}
\description{
Get the path of the output directory. This is the folder where all the 
outputs are.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
