% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\name{getPC}
\alias{getPC}
\alias{getPC,Tsne-method}
\alias{getPC,getPC-method}
\title{Getter of the PC slot of Tsne class}
\usage{
getPC(theObject)

\S4method{getPC}{Tsne}(theObject)
}
\arguments{
\item{theObject}{Tsne object}
}
\value{
PC of the Tsne object
}
\description{
Get the PC from the Tsne object
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
