% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getPCs}
\alias{getPCs}
\title{Getter of the PCs slot of scRNAseq class}
\usage{
getPCs(theObject)
}
\arguments{
\item{theObject}{scRNAseq object}
}
\value{
PCs
}
\description{
Get the vector of PCs.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
