% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getPerplexities}
\alias{getPerplexities}
\title{Getter of the perplexities slot of scRNAseq class}
\usage{
getPerplexities(theObject)
}
\arguments{
\item{theObject}{scRNAseq object}
}
\value{
Vector of perplexities
}
\description{
Get the vector of perplexities
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
