% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\name{getPerplexity}
\alias{getPerplexity}
\alias{getPerplexity,Tsne-method}
\alias{getPerplexity,getPerplexity-method}
\title{Getter of the perplexity slot of Tsne class}
\usage{
getPerplexity(theObject)

\S4method{getPerplexity}{Tsne}(theObject)
}
\arguments{
\item{theObject}{Tsne object}
}
\value{
perplexity of the Tsne object
}
\description{
Get the perplexity from the Tsne object
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
