% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\name{getSceNorm}
\alias{getSceNorm}
\alias{getSceNorm,scRNAseq-method}
\alias{getSceNorm,getSceNorm-method}
\title{Getter of sceNorm slot of scRNAseq class}
\usage{
getSceNorm(theObject)

\S4method{getSceNorm}{scRNAseq}(theObject)
}
\arguments{
\item{theObject}{scRNAseq object}
}
\value{
The SCE object assigned to the sceNorm slot.
}
\description{
Get the SCE object assigned to the sceNorm slot. This contains the colData,
the rowData, and the normalized count matrix.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
