% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\name{getSpecies}
\alias{getSpecies}
\alias{getSpecies,scRNAseq-method}
\alias{getSpecies,getSpecies-method}
\title{Getter of getSpecies slot of scRNAseq class}
\usage{
getSpecies(theObject)

\S4method{getSpecies}{scRNAseq}(theObject)
}
\arguments{
\item{theObject}{scRNAseq object}
}
\value{
The species assigned to the slot.
}
\description{
Get the species assigne to the slot. This is the studied species.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
