% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getters-Dbscan}
\alias{getters-Dbscan}
\alias{getName}
\alias{getEpsilon,Dbscan-method}
\alias{getEpsilon}
\alias{getMinPoints,Dbscan-method}
\alias{getMinPoints}
\alias{getClustering,Dbscan-method}
\alias{getClustering}
\title{getters-Dbscan}
\usage{
getName(theObject)

getEpsilon(theObject)

getMinPoints(theObject)

getClustering(theObject)
}
\arguments{
\item{theObject}{A Dbscan object. See description or ?Dbscan.}
}
\description{
Retrieve data of a slot of a Dbscan-class object.

getName: Get the name of the Dbscan.

getEpsilon: Get the epsilon used.

getMinPoints: Get the MinPoint used.

getClustering: Get the matrix of DBSCAN clustering.
}
