% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getters-scRNAseq}
\alias{getters-scRNAseq}
\alias{getExperimentName}
\alias{getCountMatrix,scRNAseq-method}
\alias{getCountMatrix}
\alias{getSceNorm,scRNAseq-method}
\alias{getSceNorm}
\alias{getSpecies,scRNAseq-method}
\alias{getSpecies}
\alias{getOutputDirectory,scRNAseq-method}
\alias{getOutputDirectory}
\alias{getTSNEList,scRNAseq-method}
\alias{getTSNEList}
\alias{getDbscanList,scRNAseq-method}
\alias{getDbscanList}
\alias{getCellsSimilarityMatrix,scRNAseq-method}
\alias{getCellsSimilarityMatrix}
\alias{getClustersSimilarityMatrix,scRNAseq-method}
\alias{getClustersSimilarityMatrix}
\alias{getClustersSimiliratyOrdered,scRNAseq-method}
\alias{getClustersSimiliratyOrdered}
\alias{getMarkerGenesList,scRNAseq-method}
\alias{getMarkerGenesList}
\alias{getClustersMarkers,scRNAseq-method}
\alias{getClustersMarkers}
\alias{getGenesInfos,scRNAseq-method}
\alias{getGenesInfos}
\title{getters-scRNAseq}
\usage{
getExperimentName(theObject)

getCountMatrix(theObject)

getSceNorm(theObject)

getSpecies(theObject)

getOutputDirectory(theObject)

getTSNEList(theObject)

getDbscanList(theObject)

getCellsSimilarityMatrix(theObject)

getClustersSimilarityMatrix(theObject)

getClustersSimiliratyOrdered(theObject)

getMarkerGenesList(theObject)

getClustersMarkers(theObject)

getGenesInfos(theObject)
}
\arguments{
\item{theObject}{A scRNA-seq object. See description or ?scRNAseq.}
}
\description{
Retrieve data of a slot of a scRNA-seq object.

getExperimentName: Get the name of the experiment.

getCountMatrix: Get the count matrix.

getSceNorm: Get the SingleCellExperiment object used.

getSpecies: Get the species.

getOutputDirectory: Get the path of the output directory.

getTSNEList: Get the list of Tsne objects.

getDbscanList: Get the list of Dbscan objects.

getCellsSimilarityMatrix: Get the cell similarity matrix.

getClustersSimilarityMatrix: Get the cluster similarity matrix.

getClustersSimiliratyOrdered: Get the clusters ordered by similarity.

getMarkerGenesList: Get the list of marker genes by clusters.

getClustersMarkers: Get the most significant markers by clusters.

getGenesInfos: Get informations about marker genes.
}
