% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getters-Tsne}
\alias{getters-Tsne}
\alias{getName}
\alias{getPerplexity,Tsne-method}
\alias{getPerplexity}
\alias{getPC,Tsne-method}
\alias{getPC}
\alias{getCoordinates,Tsne-method}
\alias{getCoordinates}
\title{getters-Tsne}
\usage{
getName(theObject)

getPerplexity(theObject)

getPC(theObject)

getCoordinates(theObject)
}
\arguments{
\item{theObject}{A Tsne object. See description or ?Tsne.}
}
\description{
Retrieve the data of a slot of a Tsne-class object.

getName: Get the name of the tSNE.

getPerplexity: Get the perplexity used.

getPC: Get the PC used.

getCoordinates: Get the matrix of tSNE coordinates.
}
