% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getters}
\alias{getters}
\alias{getExperimentName,scRNAseq-method}
\alias{getExperimentName}
\alias{getCountMatrix,scRNAseq-method}
\alias{getCountMatrix}
\alias{getSceNorm,scRNAseq-method}
\alias{getSceNorm}
\alias{getSpecies,scRNAseq-method}
\alias{getSpecies}
\alias{getOutputDirectory,scRNAseq-method}
\alias{getOutputDirectory}
\alias{getTSNEList,scRNAseq-method}
\alias{getTSNEList}
\alias{getDbscanList,scRNAseq-method}
\alias{getDbscanList}
\alias{getCellsSimilarityMatrix,scRNAseq-method}
\alias{getCellsSimilarityMatrix}
\alias{getClustersSimilarityMatrix,scRNAseq-method}
\alias{getClustersSimilarityMatrix}
\alias{getClustersSimiliratyOrdered,scRNAseq-method}
\alias{getClustersSimiliratyOrdered}
\alias{getMarkerGenesList,scRNAseq-method}
\alias{getMarkerGenesList}
\alias{getClustersMarkers,scRNAseq-method}
\alias{getClustersMarkers}
\alias{getGenesInfos,scRNAseq-method}
\alias{getGenesInfos}
\alias{getName,Tsne-method}
\alias{getPerplexity,Tsne-method}
\alias{getPerplexity}
\alias{getPC,Tsne-method}
\alias{getPC}
\alias{getCoordinates,Tsne-method}
\alias{getCoordinates}
\alias{getName,Dbscan-method}
\alias{getName}
\alias{getEpsilon,Dbscan-method}
\alias{getEpsilon}
\alias{getMinPoints,Dbscan-method}
\alias{getMinPoints}
\alias{getClustering,Dbscan-method}
\alias{getClustering}
\title{getters}
\usage{
getExperimentName(theObject)

getCountMatrix(theObject)

getSceNorm(theObject)

getSpecies(theObject)

getOutputDirectory(theObject)

getTSNEList(theObject)

getDbscanList(theObject)

getCellsSimilarityMatrix(theObject)

getClustersSimilarityMatrix(theObject)

getClustersSimiliratyOrdered(theObject)

getMarkerGenesList(theObject)

getClustersMarkers(theObject)

getGenesInfos(theObject)

\S4method{getName}{Tsne}(theObject)

getPerplexity(theObject)

getPC(theObject)

getCoordinates(theObject)

getName(theObject)

getEpsilon(theObject)

getMinPoints(theObject)

getClustering(theObject)
}
\arguments{
\item{theObject}{A scRNA-seq object. See description or ?scRNAseq.}
}
\value{
getExperimentName: Get the name of the experiment.

getCountMatrix: Get the count matrix.

getSceNorm: Get the SingleCellExperiment object used.

getSpecies: Get the species.

getOutputDirectory: Get the path of the output directory.

getTSNEList: Get the list of Tsne objects.

getDbscanList: Get the list of Dbscan objects.

getCellsSimilarityMatrix: Get the cell similarity matrix.

getClustersSimilarityMatrix: Get the cluster similarity matrix.

getClustersSimiliratyOrdered: Get the clusters ordered by similarity.

getMarkerGenesList: Get the list of marker genes by clusters.

getClustersMarkers: Get the most significant markers by clusters.

getGenesInfos: Get informations about marker genes.

getPerplexity: Get the perplexity used.

getPC: Get the PC used.

getCoordinates: Get the matrix of tSNE coordinates.

getName: Get the name of the tSNE or Dbscan object.

getEpsilon: Get the epsilon used.

getMinPoints: Get the MinPoint used.

getClustering: Get the matrix of DBSCAN clustering.
}
\description{
Retrieve data of a slot of a scRNA-seq, Tsne or Dbscan object.
}
\examples{
## Load the count matrix
countmatrixPath <- system.file("extdata/test_countMatrix.tsv", 
                            package="conclus")
countMatrix <- loadCountMatrix(file=countmatrixPath, header=TRUE, dec=".",
                                sep='\t')

## Load the coldata
coldataPath <- system.file("extdata/test_colData_filtered.tsv", 
                            package="conclus")
columnsMetaData <- loadColdata(file=coldataPath, columnCell="cell_ID",
                                header=TRUE, dec=".", sep='\t')

## Create the initial object
scr <- singlecellRNAseq(experimentName = "Bergiers",
                countMatrix     = countMatrix,
                species         = "mouse",
                outputDirectory = "YourOutputDirectory")
                
experimentName <- getExperimentName(scr)
countMatrix <- getCountMatrix(scr)
species <- getSpecies(scr)
outputDirectory <- getOutputDirectory(scr)

}
