% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\name{getTSNEList}
\alias{getTSNEList}
\alias{getTSNEList,scRNAseq-method}
\alias{getTSNEList,getTSNEList-method}
\title{Getter of the tSNEList slot of scRNAseq class}
\usage{
getTSNEList(theObject)

\S4method{getTSNEList}{scRNAseq}(theObject)
}
\arguments{
\item{theObject}{scRNAseq object}
}
\value{
List of tSNE results
}
\description{
Get a list of tSNE results
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
