% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\name{getTsneName}
\alias{getTsneName}
\alias{getTsneName,Tsne-method}
\alias{getTsneName,getTsneName-method}
\title{Getter of the name slot of Tsne class}
\usage{
getTsneName(theObject)

\S4method{getTsneName}{Tsne}(theObject)
}
\arguments{
\item{theObject}{Tsne object}
}
\value{
Name of the Tsne object
}
\description{
Get the name from the Tsne object
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
