% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_and_clustering_functions.R
\name{initialisePath}
\alias{initialisePath}
\title{Create all needed directories for CONCLUS output.}
\usage{
initialisePath(dataDirectory)
}
\arguments{
\item{dataDirectory}{output directory for a given CONCLUS run (supposed to be the same for one experiment during the workflow).}
}
\description{
Create all needed directories for CONCLUS output.
}
