% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataset.R
\name{loadColdata}
\alias{loadColdata}
\title{loadColdata}
\usage{
loadColdata(file, columnCells, header, sep, dec)
}
\arguments{
\item{file}{Path of the coldata to load.}

\item{columnCells}{Column name containing cell names/id.}

\item{header}{Set TRUE if the first row of the table corresponds to the
column names, and FALSE if it doesn't.}

\item{sep}{Character used in the table to separate the fields. 
Usually it's ' ' ';' ',' or '\\t'.}

\item{dec}{Character used in the table for decimal points.
Usually '.' or ',' .}
}
\value{
The coldata with cellName column
}
\description{
This function allows to import the coldata. All parameters are mandatory 
and no default values, so it is necessary to know how is the coldata before
importing. This function also replaces via 'columnCells' parameter the 
name of the column containing cell names by 'cellName' for the 
good walk of CONCLUS.
}
\examples{
coldataPath <- system.file("extdata/test_colData_filtered.tsv", 
                            package = "conclus")
                        
loadColdata(file=coldataPath, columnCells="cell_ID", header=TRUE, 
            sep='\t', dec=".")

}
