% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataset.R
\name{loadDataOrMatrix}
\alias{loadDataOrMatrix}
\title{loadDataOrMatrix}
\usage{
loadDataOrMatrix(file, type, columnID, header=TRUE, sep='\t', dec=".")
}
\arguments{
\item{file}{Path to the rowData, colData or Matrix.}

\item{type}{Values should be "coldata", "rowdata", or "countMatrix".}

\item{columnID}{For raw amd col data, column name containing cells/genes 
names/id.}

\item{header}{Set TRUE if the first row of the table corresponds to the
column names, and FALSE if it doesn't. Default=TRUE.}

\item{sep}{Character used in the table to separate the fields. 
Usually it's ' ' ';' ',' or '\\t'. Default='\\t'.}

\item{dec}{Character used in the table for decimal points.
Usually '.' or ',' . Default=".".}
}
\value{
The formatted row, col data or the matrix.
}
\description{
This function allows to import the coldata, rowData or the countMatrix. It 
formats each type of data to follow the requirements of CONCLUS.
}
\examples{

## ColData
coldataPath <- system.file("extdata/test_colData_filtered.tsv", 
                            package = "conclus")                 
loadColdata(file=coldataPath, type="coldata", columnID="cell_ID")

## RowData
rowdataPath <- system.file("extdata/test_rowData_filtered.tsv",
                            package="conclus")
loadRowdata(file=rowdataPath, type="rowdata", columnID="gene_ID")

## CountMatrix
countmatrixPath <- system.file("extdata/Bergiers_counts_matrix_filtered.tsv",
                                package = "conclus")
loadCountMatrix(file=countmatrixPath, type="countMatrix")

}
