% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataset.R
\name{loadRowdata}
\alias{loadRowdata}
\title{loadRowdata}
\usage{
loadRowdata(file, columnGenes, header, sep, dec)
}
\arguments{
\item{file}{Path of the rowdata. to load.}

\item{columnGenes}{Column name or indice of the column containing 
genes SYMBOL.}

\item{header}{Set TRUE if the first row of the table corresponds to the
column names, and FALSE if it doesn't.}

\item{sep}{Character used in the table to separate the fields. 
Usually it's ' ' ';' ',' or '\\t'.}

\item{dec}{Character used in the table for decimal points.
Usually '.' or ',' .}
}
\value{
The rowdata with nameInCountMatrix column
}
\description{
This function allows to import the rowdata. All parameters are mandatory 
and no default values, so it is necessary to know how is the rowdata before 
importing. This function also replaces via 'columnGenes' parameter the 
name of the column containing gene names by 'nameInCountMatrix' for the 
good walk of CONCLUS.
}
\examples{
rowdataPath <- system.file("extdata/test_rowData_filtered.tsv",
                            package="conclus")
                        
loadRowdata(file=rowdataPath, columnGenes="gene_ID", header=TRUE, 
            sep='\t', dec=".")

}
