% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{membership}
\alias{membership}
\title{Get cluster memeberships from a conclus object}
\usage{
membership(x, k = NULL)
}
\arguments{
\item{x}{An object of class 'conclus', as returned by \code{conclus}.}

\item{k}{An integer between 2 and the maximum number of clusters considered by
\code{conclus}. Defaults to \code{k=NULL} and the cluster memberships for all
numbers of clusters considered is returned.}
}
\description{
Get cluster memeberships from a conclus object
}
