% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-normalization.R
\name{normaliseCountMatrix,scRNAseq-method}
\alias{normaliseCountMatrix,scRNAseq-method}
\alias{normaliseCountMatrix}
\title{normaliseCountMatrix}
\usage{
\S4method{normaliseCountMatrix}{scRNAseq}(
  theObject,
  sizes = c(20, 40, 60, 80, 100),
  rowdata = NULL,
  coldata = NULL,
  alreadyCellFiltered = FALSE,
  runQuickCluster = TRUE
)
}
\arguments{
\item{theObject}{A scRNAseq object}

\item{sizes}{Vector of size factors from scran::computeSumFactors() function.}

\item{rowdata}{Data frame containing genes informations. Default is NULL.}

\item{coldata}{Data frame containing cells informations. Default is NULL.}

\item{alreadyCellFiltered}{Logical. If TRUE, quality check and
filtering will not be applied.}

\item{runQuickCluster}{Logical. If TRUE scran::quickCluster() function will
be applied. It usually improves the normalization for medium-size count
matrices. However, it is not recommended for datasets with less than 200
cells and may take too long for datasets with more than 10000 cells.}
}
\value{
Returns a scRNASeq object with its SCEnorm slot updated. This slot
contains a SingleCellExperiment object having the normalized count matrix,
the colData (table with cells informations), and the rowData (table with the
genes informations). See ?SingleCellExperiment for more details.
}
\description{
This function uses coldata (cells informations) and rowdata (genes 
informations) to filter the count matrix. It also normalizes by using 
deconvolution with size factors.
}
\details{
This function uses the normalization method of the scater package. For more
details about the normalization used see ?scater::normalize. The size factors
 used in the normalization are calculated with scran::computeSumFactors.
}
\examples{
experimentName <- "Bergiers"
countMatrix <- matrix(sample(seq_len(40), size=4000, replace = TRUE), 
nrow=20, ncol=200)
outputDirectory <- "./"
columnsMetaData <- read.delim(
file.path("extdata/Bergiers_colData_filtered.tsv"))

scr <- scRNAseq(experimentName = experimentName, 
                countMatrix     = countMatrix, 
                species         = "mouse",
                outputDirectory = outputDirectory)

scrNorm <- normaliseCountMatrix(scr, coldata = columnsMetaData)

}
\author{
Ilyess RACHEDI, based on code by Polina PAVLOVICH and Nicolas 
DESCOSTES.
}
