% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conclus.R
\name{pamCons}
\alias{pamCons}
\alias{fannyCons}
\alias{hclustCons}
\title{Simple wrappers to clustering functions that return just the class memberships}
\usage{
pamCons(x, k)
fannyCons(x, k)
hclustCons(x, k)
}
\arguments{
\item{x}{A dissimilarity matrix as produced, for example, by \code{cluster::daisy}
or \code{stats::dist}.}

\item{k}{The number of clusters to find.}
}
\description{
Simple wrappers to clustering functions that return just the class memberships
}
\details{
The \code{conclus} function takes an argument \code{cluster} that should
  be a function with precisely 2 arguments, \code{x} and \code{k}. The
  function should return an integer vector of cluster memberships for each item
  in \code{x}.
}
