% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{plotCellSimilarity,scRNAseq-method}
\alias{plotCellSimilarity,scRNAseq-method}
\alias{plotCellSimilarity}
\title{plotCellSimilarity}
\usage{
plotCellSimilarity(theObject, colorPalette="default", 
			statePalette="default", clusteringMethod="ward.D2", 
			orderClusters=FALSE, savePlot=FALSE, plotPDF=TRUE, returnPlot=FALSE,
			width=7, height=6, onefile=FALSE, showRowNames=FALSE, 
			showColnames=FALSE, fontsize=7.5, fontsizeRow=0.03, widthPNG=800, 
			heightPNG=750, silentPlot=FALSE)
}
\arguments{
\item{theObject}{An Object of class scRNASeq for which the count matrix was
normalized (see ?normaliseCountMatrix), tSNE were calculated (see 
?generateTSNECoordinates), dbScan was run (see ?runDBSCAN), cells were
clustered (see ?clusterCellsInternal), as clusters themselves (see 
?calculateClustersSimilarity).}

\item{colorPalette}{A vector of colors for clusters. Default = "default",
see details.}

\item{statePalette}{A vector of colors for states or conditions. See details.}

\item{clusteringMethod}{Clustering method passed to hclust() function. 
See ?hclust for a list of method. Default = "ward.D2"}

\item{orderClusters}{If TRUE, clusters in the similarity matrix of cells will
be ordered by name. Default = FALSE.}

\item{savePlot}{If TRUE, the heatmap is saved in the directory defined in 
theObject (?getOutputDirectory) and in the sub-directory "pictures".}

\item{plotPDF}{If TRUE export heatmap in pdf format, if FALSE export it in 
png format. Default=TRUE.}

\item{returnPlot}{Boolean indicating if the pHeatmap object should  be
returned by the function. Default = TRUE.}

\item{width}{Width of the plot in the pdf file. See ?pdf for more details.
Default = 7.}

\item{height}{Height of the plot in the pdf file. See ?pdf for more details.
Default = 6.}

\item{onefile}{Logical: if TRUE allow multiple figures in one file. If FALSE,
generate a file with name containing the page number for each page.
Defaults to FALSE.}

\item{showRowNames}{pheatmap parameter. Boolean specifying if row names 
are displayed. Default=FALSE.}

\item{showColnames}{pheatmap parameter. Boolean specifying if column names
are displayed. Default=FALSE.}

\item{fontsize}{pheatmap parameter. Base fontsize for the plot. Default=7.5.}

\item{fontsizeRow}{pheatmap parameter. Fontsize for rownames. Default=0.03.}

\item{widthPNG}{Width of the png. See ?png for details. Default=800.}

\item{heightPNG}{Height of the png. See ?png for details. Default=750.}

\item{silentPlot}{If TRUE, does not plot the heatmap. Default=FALSE.}
}
\value{
A pheatmap object of the similarity heatmap if returnPlot is TRUE.
}
\description{
This function plots similarity matrix as a heatmap, so one can 
see similarity between parts of different clusters.
}
\details{
colorPalette/statePalette -- A vector of colors for clusters/states or 
'default' value. If 'default' is selected, the number of clusters is limited 
to 16. If an error message is thrown, re-run the function with your own color
 vector.
}
\examples{
experimentName <- "Bergiers"
countMatrix <- as.matrix(read.delim(system.file(
"tests/testthat/test_data/test_countMatrix.tsv", package="conclus")))
outputDirectory <- "YourOutputDirectory"
columnsMetaData <- read.delim(
system.file("extdata/Bergiers_colData_filtered.tsv", package="conclus"))

## Create the initial object
scr <- scRNAseq(experimentName = experimentName, 
                countMatrix     = countMatrix, 
                species         = "mouse",
                outputDirectory = outputDirectory)

## Normalize and filter the raw counts matrix
scrNorm <- normaliseCountMatrix(scr, coldata = columnsMetaData)

## Compute the tSNE coordinates
scrTsne <- generateTSNECoordinates(scrNorm, cores=5)

## Perform the clustering with dbScan
scrDbscan <- runDBSCAN(scrTsne, cores=5)

## Compute the cell similarity matrix
scrCCI <- clusterCellsInternal(scrDbscan, clusterNumber=10, cores=4)

## Calculate clusters similarity
scrCSM <- calculateClustersSimilarity(scrCCI)

## Plot the heatmap of the similarity matrix
plotCellSimilarity(scrCSM)

}
\seealso{
calculateClustersSimilarity  plotClusteredTSNE plotCellHeatmap
plotGeneExpression plotClustersSimilarity
}
\author{
Ilyess RACHEDI, based on code by Polina PAVLOVICH and Nicolas DESCOSTES.
}
