% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_and_clustering_functions.R
\name{plotCellSimilarity}
\alias{plotCellSimilarity}
\title{Save a cells similarity matrix.}
\usage{
plotCellSimilarity(sceObject, cellsSimilarityMatrix, dataDirectory,
  experimentName, colorPalette = "default", statePalette = "default",
  clusteringMethod = "ward.D2", orderClusters = FALSE,
  plotPDF = TRUE, returnPlot = FALSE, width = 7, height = 6, ...)
}
\arguments{
\item{sceObject}{a SingleCellExperiment object with your experiment.}

\item{cellsSimilarityMatrix}{an output matrix from the conclus::clusterCellsInternal() function.}

\item{dataDirectory}{output directory for CONCLUS (supposed to be the same for one experiment during the workflow).}

\item{experimentName}{name of the experiment which will appear in filenames (supposed to be the same for one experiment during the workflow).}

\item{colorPalette}{"default" or a vector of colors for the column "clusters" in the colData, for example c("yellow", "#CC79A7").}

\item{statePalette}{"default" or a vector of colors for the column "state" in the colData, for example c("yellow", "#CC79A7").}

\item{clusteringMethod}{a clustering methods passed to hclust() function.}

\item{orderClusters}{boolean, order clusters or not.}

\item{plotPDF}{if TRUE export to pdf, if FALSE export to png. 
FALSE is recommended for datasets with more than 2500 cells due to large pdf file size.}

\item{returnPlot}{boolean, return plot or not. Default if FALSE.}

\item{width}{plot width.}

\item{height}{plot height.}

\item{...}{other parameters of pdf(), pheatmap() and png() functions.}
}
\value{
A ggplot object or nothing (depends on the returnPlot parameter).
It saves the pdf in "dataDirectory/pictures" folder.
}
\description{
This function plots similarity matrix as a heatmap, so one can see similarity between parts of different clusters.
}
