% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_and_clustering_functions.R
\name{plotClusteredTSNE}
\alias{plotClusteredTSNE}
\title{Plot t-SNE. Addtionally, it can highlight clusters or states.}
\usage{
plotClusteredTSNE(sceObject, dataDirectory, experimentName,
  tSNEresExp = "", colorPalette = "default", PCs = c(4, 6, 8, 10, 20,
  40, 50), perplexities = c(30, 40), columnName = "clusters",
  returnPlot = FALSE, width = 6, height = 5, ...)
}
\arguments{
\item{sceObject}{a SingleCellExperiment object with your experiment.}

\item{dataDirectory}{output directory for CONCLUS (supposed to be the same for one experiment during the workflow).}

\item{experimentName}{name of the experiment which will appear in filenames (supposed to be the same for one experiment during the workflow).}

\item{tSNEresExp}{if t-SNE coordinates were generated in a different CONCLUS run, you can use them without renaming the files.
Please copy tsnes folder from the source run to the current one and write that experimentName in the tSNEresExp argument.}

\item{colorPalette}{"default" or a vector of colors for the column "clusters" in the colData, for example c("yellow", "#CC79A7").}

\item{PCs}{vector of PCs (will be specified in filenames).}

\item{perplexities}{vector of perplexities (will be specified in filenames).}

\item{columnName}{name of the column to plot on t-SNE dimensions.}

\item{returnPlot}{boolean, return plot or not.}

\item{width}{plot width.}

\item{height}{plot height.}

\item{...}{other arguments of the pdf() function.}
}
\value{
A ggplot object or nothing (depends on the returnPlot parameter).
}
\description{
Plot t-SNE. Addtionally, it can highlight clusters or states.
}
