% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_and_clustering_functions.R
\name{plotGeneExpression}
\alias{plotGeneExpression}
\title{plotGeneExpression}
\usage{
plotGeneExpression(geneName, experimentName, dataDirectory,
  graphsDirectory = "pictures", sceObject, tSNEpicture = 1,
  commentName = "", palette = c("grey", "red", "#7a0f09", "black"),
  returnPlot = FALSE, savePlot = TRUE, alpha = 1, limits = NA,
  pointSize = 1, width = 6, height = 5, ...)
}
\arguments{
\item{geneName}{name of the gene you want to plot.}

\item{experimentName}{name of the experiment which appears in filenames (supposed to be the same for one experiment during the workflow).}

\item{dataDirectory}{output directory for CONCLUS (supposed to be the same for one experiment during the workflow).}

\item{graphsDirectory}{name of the subdirectory where to put graphs. Default is "dataDirectory/pictures".}

\item{sceObject}{a SingleCellExperiment object with your experiment.}

\item{tSNEpicture}{number of the picture you want to use for plotting. 
Please check "dataDirectory/tsnes" or "dataDirectory/pictures/tSNE_pictures/clusters" to get the number, it is usually from 1 to 14.}

\item{commentName}{comment you want to specify in the filename.}

\item{palette}{color palette for the legend.}

\item{returnPlot}{boolean, should the function return a ggplot object or not.}

\item{savePlot}{boolean, should the function export the plot to pdf or not.}

\item{alpha}{opacity of the points of the plot.}

\item{limits}{range of the gene expression shown in the legend.
This option allows generating t-SNE plots with equal color
scale to compare the expression of different genes. By default, limits are the range
of expression of a selected gene.}

\item{pointSize}{size of the point.}

\item{width}{plot width.}

\item{height}{plot height.}

\item{...}{other parameters of the pdf() function.}
}
\value{
A ggplot object of the plot if needed.
}
\description{
The function saves a t-SNE plot colored by expression of a given gene. 
Warning: filename with t-SNE results is hardcoded, so please don't rename the output file.
}
