% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_and_clustering_functions.R
\name{rankGenes}
\alias{rankGenes}
\title{Rank marker genes by statistical significance.}
\usage{
rankGenes(sceObject, clustersSimilarityMatrix, dataDirectory,
  experimentName, column = "clusters")
}
\arguments{
\item{sceObject}{a SingleCellExperiment object with your experiment.}

\item{clustersSimilarityMatrix}{matrix, result of conclus::calculateClustersSimilarity() function.}

\item{dataDirectory}{output directory for CONCLUS (supposed to be the same for one experiment during the workflow).}

\item{experimentName}{name of the experiment which will appear in filenames (supposed to be the same for one experiment during the workflow).}

\item{column}{name of the column with a clustering result.}
}
\description{
This function searches marker genes for each cluster. It saves tables in the "dataDirectory/marker_genes" directory,
one table per cluster.
}
