% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{representatives}
\alias{representatives}
\title{Get representative items from a consensus clustering}
\usage{
representatives(x, k)
}
\arguments{
\item{x}{The output of a call to \code{conclus}.}

\item{k}{The number of clusters you believe there to be.}
}
\description{
Get representative items from a consensus clustering
}
\details{
The function returns the rownames of the representative items, as
  judged by them having the highest item consensus within their allocated
  cluster. If there are ties, the first item is arbitrarily selected.
}
\seealso{
\code{\link{consensus}}, \code{\link{conclus}}, \code{\link{assignClusters}}
}
