% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rscale}
\alias{rscale}
\title{Robustly scale a numeric vector.}
\usage{
rscale(x, center = TRUE, scale = TRUE, na.action = na.omit)
}
\arguments{
\item{x}{A numeric vector.}

\item{center}{Logical, whether to center using the median. Defauts to
\code{center = TRUE}.}

\item{scale}{Logical, whether to scale using the mad. Defaults to
\code{scale = TRUE}.}

\item{na.action}{What to do when there are missing values. Defaults to
\code{na.action=na.omit} and missing values are returned in the same
places as in \code{x}. The alternative is \code{na.action=na.fail} which
will cause the funtion to fail if there are missing values.}
}
\description{
Robustly scale a numeric vector.
}
\details{
The vector will be centered by its median and then scaled by its
  median absolute deviation. There may be better (more efficient) ways of
  doing this (say by using M-estimators of location and scale), but for most
  cases, this approach ought to be fine.
}
