% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_and_clustering_functions.R
\name{runClustering}
\alias{runClustering}
\title{DBSCAN clustering on t-SNE results.}
\usage{
runClustering(tSNEResults, sceObject, dataDirectory, experimentName,
  epsilon = c(1.3, 1.4, 1.5), minPoints = c(3, 4), k = 0,
  deepSplit = 4, clusteringMethod = "ward.D2", cores = 14,
  deleteOutliers = TRUE, PCs = c(4, 6, 8, 10, 20, 40, 50),
  perplexities = c(30, 40), randomSeed = 42)
}
\arguments{
\item{tSNEResults}{the result of conclus::generateTSNECoordinates() function.}

\item{sceObject}{a SingleCellExperiment object with your experiment.}

\item{dataDirectory}{output directory of a given CONCLUS run (supposed to be the same for one experiment during the workflow).}

\item{experimentName}{name of the experiment which appears in filenames (supposed to be the same for one experiment during the workflow).}

\item{epsilon}{a parameter of fpc::dbscan() function.}

\item{minPoints}{a parameter of fpc::dbscan() function.}

\item{k}{preferred number of clusters. Alternative to deepSplit.}

\item{deepSplit}{intuitive level of clustering depth. Options are 1, 2, 3, 4.}

\item{clusteringMethod}{a clustering methods passed to hclust() function.}

\item{cores}{maximum number of jobs that CONCLUS can run in parallel.}

\item{deleteOutliers}{Whether cells which were often defined as outliers by dbscan must be deleted.
It will require recalculating of the similarity matrix of cells. Default is FALSE.
Usually those cells appear in an "outlier" cluster and can be easier distinguished and deleted later
if necessary.}

\item{PCs}{a vector of first principal components.
For example, to take ranges 1:5 and 1:10 write c(5, 10).}

\item{perplexities}{a vector of perplexity for t-SNE.}

\item{randomSeed}{random seed for reproducibility.}
}
\value{
A list containing filtered from outliers SingleCellExperiment object and cells similarity matrix.
}
\description{
This function provides consensus DBSCAN clustering based on the results of t-SNE. 
You can tune algorithm parameters in options to get the number of clusters you want.
}
