% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_and_clustering_functions.R
\name{saveGenesInfo}
\alias{saveGenesInfo}
\title{Save gene information into a table or tables for multiple inputs.}
\usage{
saveGenesInfo(dataDirectory = "", inputDir = "", outputDir = "",
  pattern = "", databaseDir = system.file("extdata", package =
  "conclus"), sep = ";", header = TRUE, startFromFile = 1,
  groupBy = "clusters", orderGenes = "initial", getUniprot = TRUE,
  silent = FALSE, coresGenes = 20)
}
\arguments{
\item{dataDirectory}{a directory with CONCLUS output. You can specify either 
dataDirectory, then inputDir and outputDir will be hardcoded, or inputDir and outputDir only.
The first is recommended during running CONCLUS workflow when the second option
is comfortable when you created input tables with genes manually.}

\item{inputDir}{input directory containing text files. These files can be obtained by 
applying conclus::saveMarkersLists() function or created manually. Each file must be a 
data frame with the first column called "geneName" containing gene symbols and (or) ENSEMBL IDs.}

\item{outputDir}{output directory.}

\item{pattern}{a pattern of file names to take.}

\item{databaseDir}{a path to the database "Mmus_gene_database_secretedMol.tsv". It is provided with the conclus package.}

\item{sep}{a parameter of read.delim() function.}

\item{header}{whether or not your input files have a header.}

\item{startFromFile}{number of the input file to start with. The function approaches files one by one.
It uses web scraping method to collect publicly available info from MGI, NCBI and UniProt websites.
Sometimes, if the Internet connection is not reliable, the function can drop. 
In this case, it is comfortable to start from the failed file and not to redo the previous ones.}

\item{groupBy}{a column in the input table used for grouping the genes in the output tables.}

\item{orderGenes}{if "initial" then the order of genes will not be changed.}

\item{getUniprot}{boolean, whether to get information from UniProt or not. Default is TRUE.
Sometimes, the connection to the website is not reliable. 
If you tried a couple of times and it failed, select FALSE.}

\item{silent}{whether to show messages from intermediate steps or not.}

\item{coresGenes}{maximum number of jobs that the function can run in parallel.}
}
\value{
It saves text files either in the 'dataDirectory/marker_genes/saveGenesInfo' or outputDir 
depending on whether you specify dataDirectory or (inpitDir and outputDir) explicitly.
}
\description{
This function runs conclus::getGenesInfo() function for all tables into the inputDir 
and saves the result into the outputDir.
}
