% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{scRNAseq-class}
\alias{scRNAseq-class}
\alias{scRNAseq}
\title{The scRNAseq class}
\description{
S4 class and the main class used by CONCLUS containing the different steps
to analyse rare cell populations.
}
\section{Slots}{

\describe{
   \item{\code{experimentName}:}{Class \code{"character"},
   It's the name of the experiment. Many output of scRNAseq will use
   this name.}
   \item{\code{countMatrix}:}{Integer of class \code{"matrix"},
   it's the raw matrix with reads or unique molecular identifiers (UMIs).}
   \item{\code{sceNorm}:}{Object of class SingleCellExperiment.
   Slot empty by default. Contain the colData giving informations about
   cell; the rowData giving informations about genes and the normalized
   count matrix. Fill it with normaliseCountMatrix method.}
   \item{\code{species}:}{ Class \code{"character"},
   it's the species of interest. Actually it's limited to "mouse" or
   "human".}
   \item{\code{outputDirectory}:}{Character of class \code{"character"},
   it's the path where the outputs have to go.}
   \item{\code{tSNEList}:}{List of objects of class \code{"Tsne"},
   It's the name of the experiment. Many output of scRNAseq will use
   this name.}
   \item{\code{dbscanList}:}{List of bjects of class \code{"Dbscan"}}
   \item{\code{cellsSimilarityMatrix}:}{ Similarity  cells * cells. Define
   how many times twos cell have been seen together across the 84 solutions
   of clustering }
   \item{\code{clustersSimilarityMatrix}:}{Character of class \code{"character"},
   it's the path where the outputs have to go.}
   \item{\code{clustersSimiliratyOrdered}:}{Object of class SingleCellExperiment.
   Slot empty by default. Contain the colData giving informations about
   cell; the rowData giving informations about genes and the normalized
   count matrix. Fill it with normaliseCountMatrix method.}
   \item{\code{markerGenesList}:}{List of data frames. Each data frame contains
   the ranked genes of one cluster.}
   \item{\code{clustersMarkers}:}{Data.frame. Contain top 10 marker genes of
   each clusters.}
   \item{\code{genesInfos}:}{Data.frame. Contain informations of the markers
    genes each clusters.}

   }
}

