% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setters.R
\name{setCellsSimilarityMatrix<-}
\alias{setCellsSimilarityMatrix<-}
\alias{setCellsSimilarityMatrix<-,scRNAseq-method}
\alias{setCellsSimilarityMatrix<-,setCellsSimilarityMatrix-method}
\title{Setter of cellsSimilarityMatrix slot of scRNAseq class}
\usage{
setCellsSimilarityMatrix(theObject) <- value

\S4method{setCellsSimilarityMatrix}{scRNAseq}(theObject) <- value
}
\arguments{
\item{theObject}{scRNAseq object}

\item{value}{cells Similarity Matrix}
}
\description{
Set a value to the cellsSimilarityMatrix slot.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
