% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setters.R
\name{setClustering<-}
\alias{setClustering<-}
\alias{setClustering<-,Dbscan-method}
\alias{setClustering<-,setClustering-method}
\title{Setter of clustering slot of Dbscan class}
\usage{
setClustering(theObject) <- value

\S4method{setClustering}{Dbscan}(theObject) <- value
}
\arguments{
\item{theObject}{Dbscan object}

\item{value}{clustering}
}
\description{
Set a clustering to the clustering slot.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
