% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setters.R
\name{setClustersMarkers<-}
\alias{setClustersMarkers<-}
\alias{setClustersMarkers<-,scRNAseq-method}
\alias{setClustersMarkers<-,setClustersMarkers-method}
\title{Setter of clustersMarkers slot of scRNAseq class}
\usage{
setClustersMarkers(theObject) <- value

\S4method{setClustersMarkers}{scRNAseq}(theObject) <- value
}
\arguments{
\item{theObject}{scRNAseq object}

\item{value}{Data.frame of best markers genes among all clusters.}
}
\description{
Set a value to the clustersMarkers slot.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
