% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setters.R
\name{setClustersSimilarityMatrix<-}
\alias{setClustersSimilarityMatrix<-}
\alias{setClustersSimilarityMatrix<-,scRNAseq-method}
\alias{setClustersSimilarityMatrix<-,setClustersSimilarityMatrix-method}
\title{Setter of clustersSimilarityMatrix slot of scRNAseq class}
\usage{
setClustersSimilarityMatrix(theObject) <- value

\S4method{setClustersSimilarityMatrix}{scRNAseq}(theObject) <- value
}
\arguments{
\item{theObject}{scRNAseq object}

\item{value}{clusters Similarity Matrix}
}
\description{
Set a value to the clustersSimilarityMatrix slot.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
