% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{setClustersSimiliratyOrdered<-}
\alias{setClustersSimiliratyOrdered<-}
\title{Setter of clustersSimiliratyOrdered slot of scRNAseq class}
\usage{
setClustersSimiliratyOrdered(theObject) <- value
}
\arguments{
\item{theObject}{scRNAseq object}

\item{value}{Factor. Label of clusters similiraty ordered.}
}
\description{
Set a value to the clustersSimiliratyOrdered slot.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
