% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setters.R
\name{setCoordinates<-}
\alias{setCoordinates<-}
\alias{setCoordinates<-,Tsne-method}
\alias{setCoordinates<-,setCoordinates-method}
\title{Setter of perplexity slot of Tsne class}
\usage{
setCoordinates(theObject) <- value

\S4method{setCoordinates}{Tsne}(theObject) <- value
}
\arguments{
\item{theObject}{Tsne object}

\item{value}{Coordinates (data.frame)}
}
\description{
Set a perplexity to the perplexity slot.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
