% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setters.R
\name{setCountMatrix<-}
\alias{setCountMatrix<-}
\alias{setCountMatrix<-,scRNAseq-method}
\alias{setCountMatrix<-,setCountMatrix-method}
\alias{setClustersSimiliratyOrdered<-,scRNAseq-method}
\alias{setClustersSimiliratyOrdered<-,setClustersSimiliratyOrdered-method}
\title{Setter of countMatrix slot of scRNAseq class}
\usage{
setCountMatrix(theObject) <- value

\S4method{setCountMatrix}{scRNAseq}(theObject) <- value

\S4method{setClustersSimiliratyOrdered}{scRNAseq}(theObject) <- value
}
\arguments{
\item{theObject}{scRNAseq object}

\item{value}{COunt matrix}
}
\description{
Set the the count matrix to the countMatrix slot.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
