% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setters.R
\name{setDbscanList<-}
\alias{setDbscanList<-}
\alias{setDbscanList<-,scRNAseq-method}
\alias{setDbscanList<-,setDbscanList-method}
\title{Setter of dbscanList slot of scRNAseq class}
\usage{
setDbscanList(theObject) <- value

\S4method{setDbscanList}{scRNAseq}(theObject) <- value
}
\arguments{
\item{theObject}{scRNAseq object}

\item{value}{List of DBSCAN results}
}
\description{
Set a list of DBSCAN solutions to the dbscanList slot.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
