% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setters.R
\name{setDbscanName<-}
\alias{setDbscanName<-}
\alias{setDbscanName<-,Dbscan-method}
\alias{setDbscanName-method,setName-method}
\title{Setter of name slot of Dbscan class}
\usage{
setDbscanName(theObject) <- value

\S4method{setDbscanName}{Dbscan}(theObject) <- value
}
\arguments{
\item{theObject}{Dbscan object}

\item{value}{Name}
}
\description{
Set a name to the name slot.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
