% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setters.R
\name{setEpsilon<-}
\alias{setEpsilon<-}
\alias{setEpsilon<-,Dbscan-method}
\alias{setEpsilon<-,setEpsilon-method}
\title{Setter of epsilon slot of Dbscan class}
\usage{
setEpsilon(theObject) <- value

\S4method{setEpsilon}{Dbscan}(theObject) <- value
}
\arguments{
\item{theObject}{Dbscan object}

\item{value}{epsilon}
}
\description{
Set a epsilon to the epsilon slot.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
