% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setters.R
\name{setExperimentName<-}
\alias{setExperimentName<-}
\alias{setExperimentName<-,scRNAseq-method}
\alias{setExperimentName<-,setExperimentName-method}
\title{Setter of experimentName slot of scRNAseq class}
\usage{
setExperimentName(theObject) <- value

\S4method{setExperimentName}{scRNAseq}(theObject) <- value
}
\arguments{
\item{theObject}{scRNAseq object}

\item{value}{Name of the experiment}
}
\description{
Set the experiment name to the experimentName slot. Experiment name
describes the name of th experiment of the Conclus analysis.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
