% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setters.R
\name{setGenesInfos<-}
\alias{setGenesInfos<-}
\alias{setGenesInfos<-,scRNAseq-method}
\alias{setGenesInfos<-,setGenesInfos-method}
\title{Setter of genesInfos slot of scRNAseq class}
\usage{
setGenesInfos(theObject) <- value

\S4method{setGenesInfos}{scRNAseq}(theObject) <- value
}
\arguments{
\item{theObject}{scRNAseq object}

\item{value}{Data.frame containing some informations about markers genes.}
}
\description{
Set a value to the genesInfos slot.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
