% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setters.R
\name{setMarkerGenesList<-}
\alias{setMarkerGenesList<-}
\alias{setMarkerGenesList<-,scRNAseq-method}
\alias{setMarkerGenesList<-,setMarkerGenesList-method}
\title{Setter of markerGenesList slot of scRNAseq class}
\usage{
setMarkerGenesList(theObject) <- value

\S4method{setMarkerGenesList}{scRNAseq}(theObject) <- value
}
\arguments{
\item{theObject}{scRNAseq object}

\item{value}{List of markers genes (data.frames) for each cluster.}
}
\description{
Set a value to the markerGenesList slot.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
