% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setters.R
\name{setMinPoints<-}
\alias{setMinPoints<-}
\alias{setMinPoints<-,Dbscan-method}
\alias{setMinPoints<-,setMinPoints-method}
\title{Setter of minPoints slot of Dbscan class}
\usage{
setMinPoints(theObject) <- value

\S4method{setMinPoints}{Dbscan}(theObject) <- value
}
\arguments{
\item{theObject}{Dbscan object}

\item{value}{epsilon}
}
\description{
Set a minPoints to the Dbscan slot.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
