% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setters.R
\name{setOutputDirectory<-}
\alias{setOutputDirectory<-}
\alias{setOutputDirectory<-,scRNAseq-method}
\alias{setOutputDirectory<-,setOutputDirectory-method}
\title{Setter of the outputDirectory slot of scRNAseq class}
\usage{
setOutputDirectory(theObject) <- value

\S4method{setOutputDirectory}{scRNAseq}(theObject) <- value
}
\arguments{
\item{theObject}{scRNAseq object}

\item{value}{Path of the output folder.}
}
\description{
Set the path of the output directory. This is the folder where all the 
outputs are.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
