% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setters.R
\name{setPC<-}
\alias{setPC<-}
\alias{setPC<-,Tsne-method}
\alias{setPC<-,setPC-method}
\title{Setter of PC slot of Tsne class}
\usage{
setPC(theObject) <- value

\S4method{setPC}{Tsne}(theObject) <- value
}
\arguments{
\item{theObject}{Tsne object}

\item{value}{PC}
}
\description{
Set a PC to the PC slot.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
