% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{setPCs<-}
\alias{setPCs<-}
\title{Setter of PCs slot of scRNAseq class}
\usage{
setPCs(theObject) <- value
}
\arguments{
\item{theObject}{scRNAseq object}

\item{value}{Vector of PCs}
}
\description{
Set the PCs to the PCs slot. a PC is the number of principal components
used by CONCLUS to perfom a PCA before to perform the tSNE and create the
object. By default, CONCLUS use a range of
PCs=c(8, 10, 20, 40, 50, 80, 100).
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
