% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{setPerplexities<-}
\alias{setPerplexities<-}
\title{Setter of perplexities slot of scRNAseq class}
\usage{
setPerplexities(theObject) <- value
}
\arguments{
\item{theObject}{scRNAseq object}

\item{value}{Vector of perplexities}
}
\description{
Set the perplexities to the perplexities slot.
The perplexity is a value used by tSNE algorithm.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
