% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setters.R
\name{setPerplexity<-}
\alias{setPerplexity<-}
\alias{setPerplexity<-,Tsne-method}
\alias{setPerplexity<-,setPerplexity-method}
\title{Setter of perplexity slot of Tsne class}
\usage{
setPerplexity(theObject) <- value

\S4method{setPerplexity}{Tsne}(theObject) <- value
}
\arguments{
\item{theObject}{Tsne object}

\item{value}{perplexity}
}
\description{
Set a perplexity to the perplexity slot.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
