% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setters.R
\name{setSceNorm<-}
\alias{setSceNorm<-}
\alias{setSceNorm<-,scRNAseq-method}
\alias{setSceNorm<-,setSceNorm-method}
\title{Setter of sceNorm slot of scRNAseq class}
\usage{
setSceNorm(theObject) <- value

\S4method{setSceNorm}{scRNAseq}(theObject) <- value
}
\arguments{
\item{theObject}{scRNAseq object}

\item{value}{SingleCellExperiment object with normalized count matrix}
}
\description{
Set the SCE object  to the sceNorm slot. This contains the colData,
the rowData, and the normalized count matrix.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
