% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setters.R
\name{setSpecies<-}
\alias{setSpecies<-}
\alias{setSpecies<-,scRNAseq-method}
\alias{setSpecies<-,setSpecies-method}
\title{Setter of setSpecies slot of scRNAseq class}
\usage{
setSpecies(theObject) <- value

\S4method{setSpecies}{scRNAseq}(theObject) <- value
}
\arguments{
\item{theObject}{scRNAseq object}

\item{value}{Species}
}
\description{
Set the species  to the slot. This is the studied species.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
