% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{setters-Dbscan}
\alias{setters-Dbscan}
\alias{setName<-}
\alias{setName}
\alias{setEpsilon<-,Dbscan-method}
\alias{setEpsilon<-}
\alias{setEpsilon}
\alias{setMinPoints<-,Dbscan-method}
\alias{setMinPoints<-}
\alias{setMinPoints}
\alias{setClustering<-,Dbscan-method}
\alias{setClustering<-}
\alias{setClustering}
\title{setters-Dbscan}
\usage{
setName(theObject) <- value

setEpsilon(theObject) <- value

setMinPoints(theObject) <- value

setClustering(theObject) <- value
}
\arguments{
\item{theObject}{A Dbscan object to update. See description or ?Dbscan.}

\item{value}{The value to update the slot with. See description or ?Dbscan.}
}
\description{
Update a slot of a Dbscan-class object.

setName: Update the Dbscan name slot with a character string.

setEpsilon: Update the epsilon slot with a vector of numeric.

setMinPoints: Update the minPoints slot with a vector of numeric.

setClustering: Update the clustering slot with a matrix of numeric.
}
