% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{setters-scRNAseq}
\alias{setters-scRNAseq}
\alias{setExperimentName<-}
\alias{setExperimentName}
\alias{setCountMatrix<-,scRNAseq-method}
\alias{setCountMatrix<-}
\alias{setCountMatrix}
\title{setters-scRNAseq}
\usage{
setExperimentName(theObject) <- value

setCountMatrix(theObject) <- value
}
\arguments{
\item{theObject}{A scRNA-seq object to update.}

\item{value}{The value to update the slot with.}
}
\description{
setExperimentName: Update the experiment name slot with a character string.

setCountMatrix: Update the countMatrix slot with a matrix of numeric.
}
