% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{setters-Tsne}
\alias{setters-Tsne}
\alias{setName<-}
\alias{setName}
\alias{setPC<-,Tsne-method}
\alias{setPC<-}
\alias{setPC}
\alias{setPerplexity<-,Tsne-method}
\alias{setPerplexity<-}
\alias{setPerplexity}
\alias{setCoordinates<-,Tsne-method}
\alias{setCoordinates<-}
\alias{setCoordinates}
\title{setters-Tsne}
\usage{
setName(theObject) <- value

setPC(theObject) <- value

setPerplexity(theObject) <- value

setCoordinates(theObject) <- value
}
\arguments{
\item{theObject}{A Tsne object to update. See description or ?Tsne.}

\item{value}{The value to update the slot with. See description or ?Tsne.}
}
\description{
Update a slot of a Tsne-class object.

setName: Update the Tsne name slot with a character string.

setPC: Update the pc slot with a vector of numeric.

setPerplexity: Update the perplexity slot with a vector of numeric.

setCoordinates: Update the coordinates slot with a matrix of numeric.
}
