% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{setters}
\alias{setters}
\alias{setExperimentName<-}
\alias{setCountMatrix<-,scRNAseq-method}
\alias{setCountMatrix<-}
\title{setters}
\usage{
setExperimentName(theObject) <- value

setCountMatrix(theObject) <- value
}
\arguments{
\item{theObject}{A scRNA-seq object to update.}

\item{value}{The value to update the slot with.}
}
\description{
Update the experiment name slot with value.

Update the countMatrix slot with value.
}
