% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setters.R
\name{setTSNEList<-}
\alias{setTSNEList<-}
\alias{setTSNEList<-,scRNAseq-method}
\alias{setTSNEList<-,setTSNEList-method}
\title{Setter of setTSNEList slot of scRNAseq class}
\usage{
setTSNEList(theObject) <- value

\S4method{setTSNEList}{scRNAseq}(theObject) <- value
}
\arguments{
\item{theObject}{scRNAseq object}

\item{value}{List of tSNE results}
}
\description{
Set a list of tSNE solutions to the TSNEList slot.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
