% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/setters.R
\name{setTsneName<-}
\alias{setTsneName<-}
\alias{setTsneName<-,Tsne-method}
\alias{setTsneName-method,setName-method}
\title{Setter of name slot of Tsne class}
\usage{
setTsneName(theObject) <- value

\S4method{setTsneName}{Tsne}(theObject) <- value
}
\arguments{
\item{theObject}{Tsne object}

\item{value}{Name}
}
\description{
Set a name to the name slot.
}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES, based on code by Polina
PAVLOVICH and Artem ADAMOV.
}
