% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conclus-package.R
\docType{data}
\name{stjude}
\alias{stjude}
\title{St Jude leukemia data}
\format{A data frame containing data 983 features (columns) measured on 246 patients (rows).}
\source{
http://portals.broadinstitute.org/cgi-bin/cancer/publications/view/87 Beware! the URL
  has been known to change. An alternative is to go to http://portals.broadinstitute.org/cgi-bin/cancer
  and search for the paper by Monti et al.
}
\description{
St Jude leukemia data
}
\note{
It is reported that there are 6 distinct groups within the data, so something
  close to 6 ought to be identified by clustering either the rows or the columns.
}
\references{
S. Monti, P. Tamayo, J. Mesirov and T. Golub, Consensus clustering: a
  resampling-based method for class discovery and visualization of gene expression data,
  Machine Learning, 52, 91 -- 118, 2003
}
