% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.conclus}
\alias{summary.conclus}
\title{Compute and return summaries of a 'conclus' object.}
\usage{
\method{summary}{conclus}(object, plot.it = TRUE, ...)
}
\arguments{
\item{object}{An object of class 'conclus'.}

\item{plot.it}{Whether to plot the result. Defaults to \code{plot.it=TRUE}.
If \code{plot.it=FALSE}, the output can be assigned to an object and plotted
later using \code{ggplot(object)}.}

\item{...}{Not used.}
}
\description{
Compute and return summaries of a 'conclus' object.
}
\details{
The function computes the distribution function of the elements of the
  consensus matrices returned by the call to \code{conclus}. From these, it
  then returns the ares under the curves, and the incremental proportional changes to those areas as
  the number of clusters k is increeased. It also computes the 'cluster consensus'
  which enables the user to see which of the clusters identified are the most
  stable. These statistics are intended to aid the user in selecting how many
  clusters they might reasonably believe to exist. They should be used in
  conjunction with the plot of the consensus matrices, generated via
  \code{ggplot(objec t)}, for the object that is passed to the \code{summary}
  function.
}
