% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_and_clustering_functions.R
\name{testClustering}
\alias{testClustering}
\title{To check one iteration of clustering before running full workflow CONCLUS.}
\usage{
testClustering(sceObject, dataDirectory, experimentName,
  dbscanEpsilon = 1.4, minPts = 5, perplexities = c(30),
  PCs = c(4), randomSeed = 42, width = 7, height = 7, ...)
}
\arguments{
\item{sceObject}{a SingleCellExperiment object with your experiment.}

\item{dataDirectory}{output directory (supposed to be the same for one experiment during the workflow).}

\item{experimentName}{name of the experiment which will appear in filenames (supposed to be the same for one experiment during the workflow).}

\item{dbscanEpsilon}{a parameter of fpc::dbscan() function.}

\item{minPts}{a parameter of fpc::dbscan() function.}

\item{perplexities}{vector of perplexities (t-SNE parameter).}

\item{PCs}{a vector of PCs for plotting.}

\item{randomSeed}{random seed for reproducibility.}

\item{width}{plot width.}

\item{height}{plot height.}

\item{...}{other pdf() arguments.}
}
\value{
t-SNE results, a distance graph plot, a t-SNE plot colored by test clustering solution.
}
\description{
This function generates a single clustering iteration of CONCLUS to check whether
chosen parameters for dbscan are suitable for your data.
}
