% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Tsne-class}
\alias{Tsne-class}
\alias{Tsne}
\title{The Tsne class}
\description{
S4 class. It represents the features of tSNE, like the name,
the number of PC and the perplexity used, the coordinates.
}
\section{Slots}{

 \describe{
   \item{\code{name}:}{Variable of class \code{"character"},
   It's the name of the tSNE coordinates.}
   \item{\code{pc}:}{Vector of class \code{"numeric"}. The PC is the number
   of principal components used by CONCLUS to perfom a PCA before to perform
   the tSNE and create the object. By default, CONCLUS use a range of
   PCs=c(8, 10, 20, 40, 50, 80, 100). So the PC for a Tsne object is one of
   these values.}
   \item{\code{perplexity}:}{Vector of class \code{"numeric"}. The value is
   the perplexity used by CONCLUS to perform the tSNE. By default, CONCLUS
   use a range of perplexities=c(30, 40) }
   \item{\code{coordinates}:}{Variable of class \code{"matrix"}, it contains
   the coordinates of one tSNE solution.}
   }
}

