% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtable.summary.conclus.R
\name{xtable.summary.conclus}
\alias{xtable.summary.conclus}
\title{Create an xtable from a summary.conclus object}
\usage{
\method{xtable}{summary.conclus}(x, caption = NULL, label = NULL,
  align = NULL, digits = 2, display = NULL, auto = FALSE, ...)
}
\arguments{
\item{x}{The output of \code{summary.conclus}.}

\item{caption, }{label, align, digits, display, auto, ... Arguments to \code{xtable}}
}
\description{
Create an xtable from a summary.conclus object
}
