% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-clustering.R
\name{addClustering,scRNAseq-method}
\alias{addClustering,scRNAseq-method}
\alias{addClustering}
\title{addClustering}
\usage{
addClustering(theObject, filePathAdd=NA, clusToAdd=NA)
}
\arguments{
\item{theObject}{An Object of class scRNASeq for which
?calculateClustersSimilarity was used.}

\item{filePathAdd}{Default=NA. Path to the file containing the clustering to
replace in the object. It should be made of two columns 'clusters' and
'cells'. This should be left to NA if clusToAdd is used. Default=NA.}

\item{clusToAdd}{Data frame having two columns 'clusters' and 'cells'
containing the clustering to be used in theObject. Should be left to NA if
filePathAdd is used. Default=NA.}
}
\value{
An object of class scRNASeq with its column name metadata updated.
}
\description{
This method enables to add a clustering to the existing object in order to
change the coloration of the t-sne. It is particularly useful to compare the
performance of different tools.
}
\examples{
## Load the count matrix
countmatrixPath <- system.file("extdata/test_countMatrix.tsv", 
                            package="conclus")
countMatrix <- loadDataOrMatrix(file=countmatrixPath, type="countMatrix")

## Load the coldata
coldataPath <- system.file("extdata/test_colData_filtered.tsv", 
                            package="conclus")
columnsMetaData <- loadDataOrMatrix(file=coldataPath, type="coldata", 
										columnID="cell_ID")

## Load the clusters table with the new cluster to add
clustAddTab <- read.delim(
system.file("extdata/Bergiers_clusters_table.tsv", package="conclus"))

## Create the initial object
scr <- singlecellRNAseq(experimentName = "Bergiers",
                countMatrix     = countMatrix,
                species         = "mouse",
                outputDirectory = "YourOutputDirectory")

## Normalize and filter the raw counts matrix
scrNorm <- normaliseCountMatrix(scr, coldata = columnsMetaData)

## Compute the tSNE coordinates
scrTsne <- generateTSNECoordinates(scrNorm, cores=2)

## Perform the clustering with dbScan
scrDbscan <- runDBSCAN(scrTsne, cores=2)

## Compute the cell similarity matrix
scrCCI <- clusterCellsInternal(scrDbscan, clusterNumber=10, cores=2)

## Calculate clusters similarity
scrCSM <- calculateClustersSimilarity(scrCCI)

## Ranking genes
scrS4MG <- rankGenes(scrCSM)

## Getting marker genes
scrFinal <- retrieveTopClustersMarkers(scrS4MG, removeDuplicates = FALSE)

## Getting genes info
scrInfos <- retrieveGenesInfo(scrFinal, cores=2)

## Retrieving the table indicating to which cluster each cell belongs
clustCellsDf <- retrieveTableClustersCells(scrInfos)

## Replace “10” by “9” to merge 9/10
clustCellsDf$clusters[which(clustCellsDf$clusters == 10)] <- 9

## Modifying the object to take into account the new classification
scrUpdated <- addClustering(scrInfos, clusToAdd=clustCellsDf)


}
\author{
Ilyess RACHEDI, based on code by Polina PAVLOVICH and Nicolas DESCOSTES.
}
