% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-clustering.R
\name{addClustering,scRNAseq-method}
\alias{addClustering,scRNAseq-method}
\alias{addClustering}
\title{addClustering}
\usage{
addClustering(theObject, filePathAdd=NA, clusToAdd=NA)
}
\arguments{
\item{theObject}{An Object of class scRNASeq for which 
?calculateClustersSimilarity was used.}

\item{filePathAdd}{Default=NA. Path to the file containing the clustering to 
replace in the object. It should be made of two columns 'clusters' and 
'cells'. This should be left to NA if clusToAdd is used. Default=NA.}

\item{clusToAdd}{Data frame having two columns 'clusters' and 'cells' 
containing the clustering to be used in theObject. Should be left to NA if 
filePathAdd is used. Default=NA.}
}
\value{
An object of class scRNASeq with its column name metadata updated.
}
\description{
This method enables to add a clustering to the existing object in order to 
change the coloration of the t-sne. It is particularly useful to compare the
 performance of different tools.
}
\examples{
experimentName <- "Bergiers"
countMatrix <- as.matrix(read.delim(file.path(
"tests/testthat/test_data/test_countMatrix.tsv")))
outputDirectory <- "./"
columnsMetaData <- read.delim(
file.path("extdata/Bergiers_colData_filtered.tsv"))

clustAddTab <- read.delim(
file.path("inst/extdata/Bergiers_clusters_table.tsv"))
		
## Create the initial object
scr <- scRNAseq(experimentName = experimentName, 
                countMatrix     = countMatrix, 
                species         = "mouse",
                outputDirectory = outputDirectory)

## Normalize and filter the raw counts matrix
scrNorm <- normaliseCountMatrix(scr, coldata = columnsMetaData)

## Compute the tSNE coordinates
scrTsne <- generateTSNECoordinates(scrNorm, cores=5)

## Perform the clustering with dbScan
scrDbscan <- runDBSCAN(scrTsne, cores=5)

## Compute the cell similarity matrix
scrCCI <- clusterCellsInternal(scrDbscan, clusterNumber=10, cores=4)

## Calculate clusters similarity
scrCSM <- calculateClustersSimilarity(scrCCI)

## Update clustering information
addClustering(scrCSM, clusToAdd=clustAddTab)

}
\author{
Ilyess RACHEDI, based on code by Polina PAVLOVICH and Nicolas DESCOSTES.
}
