% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-clustering.R
\name{calculateClustersSimilarity,scRNAseq-method}
\alias{calculateClustersSimilarity,scRNAseq-method}
\alias{calculateClustersSimilarity}
\title{calculateClustersSimilarity}
\usage{
calculateClustersSimilarity(theObject, clusteringMethod = "ward.D2")
}
\arguments{
\item{theObject}{An Object of class scRNASeq for which the count matrix was
normalized (see ?normaliseCountMatrix), tSNE were calculated (see 
?generateTSNECoordinates), dbScan was run (see ?runDBSCAN), and cells were
clustered (see ?clusterCellsInternal).}

\item{clusteringMethod}{Clustering method passed to hclust() function. See 
?hclust for a list of method. Default = "ward.D2".}
}
\value{
An object of class scRNASeq with its clustersSimilarityMatrix and
clustersSimiliratyOrdered slots updated.
}
\description{
Having computed cells similarity, pools information into clusters.
}
\examples{
experimentName <- "Bergiers"
countMatrix <- as.matrix(read.delim(file.path(
"tests/testthat/test_data/test_countMatrix.tsv")))
outputDirectory <- "./"
columnsMetaData <- read.delim(
file.path("extdata/Bergiers_colData_filtered.tsv"))

## Create the initial object
scr <- scRNAseq(experimentName = experimentName, 
                countMatrix     = countMatrix, 
                species         = "mouse",
                outputDirectory = outputDirectory)

## Normalize and filter the raw counts matrix
scrNorm <- normaliseCountMatrix(scr, coldata = columnsMetaData)

## Compute the tSNE coordinates
scrTsne <- generateTSNECoordinates(scrNorm, cores=5)

## Perform the clustering with dbScan
scrDbscan <- runDBSCAN(scrTsne, cores=5)

## Compute the cell similarity matrix
scrCCI <- clusterCellsInternal(scrDbscan, clusterNumber=10, cores=4)

## Calculate clusters similarity
scrCSM <- calculateClustersSimilarity(scrCCI)

}
\seealso{
plotClustersSimilarity
}
\author{
Ilyess RACHEDI, based on code by Polina PAVLOVICH and Nicolas DESCOSTES.
}
