% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-clustering.R
\name{clusterCellsInternal,scRNAseq-method}
\alias{clusterCellsInternal,scRNAseq-method}
\alias{clusterCellsInternal}
\title{clusterCellsInternal}
\usage{
clusterCellsInternal(theObject, clusterNumber=0, deepSplit=4, cores=2,
                clusteringMethod="ward.D2")
}
\arguments{
\item{theObject}{An Object of class scRNASeq for which the count matrix was
normalized (see ?normaliseCountMatrix), tSNE were calculated (see
?generateTSNECoordinates), and dbScan was run (see ?runDBSCAN),}

\item{clusterNumber}{Exact number of cluster. Default = 0 that will determine
the number of clusters automatically.}

\item{deepSplit}{Intuitive level of clustering depth. Options are 1, 2, 3, 4.
Default = 4}

\item{cores}{Maximum number of jobs that CONCLUS can run in parallel.
Default is 1.}

\item{clusteringMethod}{Clustering method passed to hclust() function. See
?hclust for a list of method. Default = "ward.D2".}
}
\value{
An object of class scRNASeq with its cellsSimilarityMatrix and sceNorm slots
updated.
}
\description{
Returns consensus clusters by using hierarchical clustering on the similarity
matrix of cells.
}
\examples{
## Object containing the results of previous steps
load(system.file("extdata/scrFull.Rdat", package="conclus"))

## Compute the cell similarity matrix
scr <- clusterCellsInternal(scr, clusterNumber=4, cores=2)

}
\seealso{
plotCellSimilarity
}
\author{
Ilyess RACHEDI, based on code by Polina PAVLOVICH and Nicolas DESCOSTES.
}
