% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{constructors}
\alias{constructors}
\alias{scRNAseq}
\title{constructors}
\usage{
scRNAseq(experimentName, countMatrix, species, outputDirectory)
}
\arguments{
\item{experimentName}{character string representing the name of the 
experiment. Many output of scRNAseq will use this name.}

\item{countMatrix}{An integer matrix representing the raw count matrix with 
reads or unique molecular identifiers (UMIs).}

\item{species}{character string representing the species of interest. 
Currently limited to "mouse" and "human".}

\item{outputDirectory}{A character string of the path to the root output 
folder.}
}
\description{
Constructor of the class scRNAseq (might be extended in the future).
}
\seealso{
scRNAseq-class
}
