% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{constructors}
\alias{constructors}
\alias{singlecellRNAseq}
\alias{TsneCluster}
\alias{DbscanCluster}
\title{constructors}
\usage{
singlecellRNAseq(experimentName, countMatrix, species, outputDirectory, 
tSNElist=list(new("Tsne")))

TsneCluster(name, pc, perplexity, coordinates)

DbscanCluster(name, epsilon, minPoints, clustering)
}
\arguments{
\item{experimentName}{character string representing the name of the
experiment. Many output of scRNAseq will use this name.}

\item{countMatrix}{An integer matrix representing the raw count matrix with
reads or unique molecular identifiers (UMIs).}

\item{species}{character string representing the species of interest.
Currently limited to "mouse" and "human".}

\item{outputDirectory}{A character string of the path to the root output
folder.}

\item{tSNElist}{List of 'Tsne' objects representing the different tSNE
coordinates generated by CONCLUS.}

\item{dbscanlist}{List of 'Dbscan' objects representing the different 
Dbscan clustering generated by CONCLUS.}

\item{cellSimMat}{A numeric Matrix defining how many times two cells have
been associated to the same cluster across the 84 solutions of clustering.}

\item{clustSimMat}{A numeric matrix comparing the robustness of the consensus
clusters.}

\item{clustSimOrdered}{A factor representing the clusters ordered by 
similarity.}

\item{markgenlist}{list of data.frames. Each data frame contains the ranked
genes of one cluster.}

\item{clustMark}{A data frame containing the top 10 (by default)
marker genes of each clusters.}

\item{name}{A 'character' string representing the name of the Dbscan
clustering.}

\item{pc}{A 'numeric' value representing the number of principal components
used by CONCLUS to perfom a PCA before calculating the tSNE.}

\item{perplexity}{A 'numeric'. Default: c(30, 40)}

\item{coordinates}{A 'numeric' matrix that contains the coordinates
of one tSNE solution.}

\item{epsilon}{A 'numeric' value. The epsilon is the distance to consider
two points belonging to the same cluster. Default = c(1.3, 1.4, 1.5)}

\item{minPoints}{A 'numeric' value. The minPoints is the minimum number
of points to construct a cluster.}

\item{clustering}{A 'matrix' that contains the result of one DBSCAN
clustering solution.}

\item{genesInfos}{A data frame containing informations of the markers
genes for each clusters.}
}
\value{
Object of class scRNAseq

Object of class Tsne

Object of class Dbscan
}
\description{
Constructor of the class scRNAseq (might be extended in the future).
}
\examples{
experimentName <- "Bergiers"
countMatrix <- as.matrix(read.delim(system.file(
"extdata/test_countMatrix.tsv", package="conclus")))
outputDirectory <- "YourOutputDirectory"
columnsMetaData <- read.delim(
system.file("extdata/Bergiers_colData_filtered.tsv", package="conclus"))

## Create the initial object
scr <- singlecellRNAseq(experimentName = experimentName,
                countMatrix     = countMatrix,
                species         = "mouse",
                outputDirectory = outputDirectory)
                

mat <- matrix(seq_len(20), ncol=2)
colnames(mat) <- c("X", "Y")
TsneCluster(name = "test", pc = 30, perplexity = 4, 
coordinates = mat)
                                             
DbscanCluster("test", 0.5, 2, matrix(1:10))
                                         
}
\seealso{
scRNAseq-class

Tsne-class

Dbscan-class
}
