% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/getters.R, R/setters.R
\docType{class}
\name{Dbscan-class}
\alias{Dbscan-class}
\alias{Dbscan}
\alias{getDbscanName}
\alias{getEpsilon}
\alias{getMinPoints}
\alias{getClustering}
\alias{setDbscanName}
\alias{setEpsilon}
\alias{setMinPoints}
\alias{setClustering}
\alias{getName,Dbscan-method}
\alias{getEpsilon,Dbscan-method}
\alias{getMinPoints,Dbscan-method}
\alias{getClustering,Dbscan-method}
\alias{setName<-,Dbscan-method}
\alias{setEpsilon<-,Dbscan-method}
\alias{setMinPoints<-,Dbscan-method}
\alias{setClustering<-,Dbscan-method}
\title{The Dbscan class}
\usage{
\S4method{getName}{Dbscan}(theObject)

\S4method{getEpsilon}{Dbscan}(theObject)

\S4method{getMinPoints}{Dbscan}(theObject)

\S4method{getClustering}{Dbscan}(theObject)

\S4method{setName}{Dbscan}(theObject) <- value

\S4method{setEpsilon}{Dbscan}(theObject) <- value

\S4method{setMinPoints}{Dbscan}(theObject) <- value

\S4method{setClustering}{Dbscan}(theObject) <- value
}
\description{
S4 class containing the features to plot DBSCAN. This constructor is internal
and is used by the method runDBSCAN.
}
\section{Functions}{
\itemize{
\item \code{getName,Dbscan-method}: Get the name of the Dbscan.

\item \code{getEpsilon,Dbscan-method}: Get the epsilon used.

\item \code{getMinPoints,Dbscan-method}: Get the MinPoint used.

\item \code{getClustering,Dbscan-method}: Get the matrix of DBSCAN clustering.

\item \code{setName<-,Dbscan-method}: Set the name of the Dbscan.

\item \code{setEpsilon<-,Dbscan-method}: Set the epsilon used.

\item \code{setMinPoints<-,Dbscan-method}: Set the minPoints used.

\item \code{setClustering<-,Dbscan-method}: Set the matrix of Dbscan clustering.
}}

\section{Slots}{

\describe{
\item{\code{name}}{A 'character' string representing the name of the Dbscan 
clustering.}

\item{\code{epsilon}}{A 'numeric' value. The epsilon is the distance to consider two 
points belonging to the same cluster. Default = c(1.3, 1.4, 1.5)}

\item{\code{minPoints}}{A 'numeric' value. The minPoints is the minimum number of 
points to construct a cluster.}

\item{\code{clustering}}{A 'matrix' that contains the result of one DBSCAN clustering
solution.}
}}

\section{Constructor}{


   Dbscan(name = "character", epsilon = "numeric", minPoints = "numeric", 
        clustering = "matrix")

   name:       Empty character string or the name of the tSNE.
   epsilon:    Empty 'numeric' representing the epsilon.
   minPoints:  Empty 'numeric' representing the minPoints value.
   clustering: Empty 'numeric' "matrix" or matrix of clustering.
}

\section{Accessors}{


    In the following snippets, x is a Dbscan object.

    getName(x): Get the name of the Dbscan.
    getEpsilon(x):    Get the epsilon used. 
    getMinPoints(x):  Get the MinPoint used.
    getClustering(x): Get the matrix of DBSCAN clustering.
}

\section{Subsetting}{


    In the following snippets, x is a Dbscan object.

    setName(x) <- value: Set the name of the Dbscan.
    setEpsilon(x) <- value:    Set the epsilon used. 
    setMinPoints(x) <- value:  Set the minPoints used.
    setClustering(x) <- value: Set the matrix of Dbscan clustering.
}

\seealso{
runDBSCAN
}
\author{
Ilyess Rachedi and Nicolas Descostes
}
