% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Dbscan-class}
\alias{Dbscan-class}
\alias{Dbscan}
\title{The Dbscan class}
\description{
S4 class containing the features to plot DBSCAN. This constructor is internal
and is used by the method runDBSCAN.
}
\section{Slots}{

\describe{
\item{\code{name}}{A 'character' string representing the name of the Dbscan
clustering.}

\item{\code{epsilon}}{A 'numeric' vector. The epsilon is the distance to consider
two points belonging to the same cluster. Default = c(1.3, 1.4, 1.5)}

\item{\code{minPoints}}{A 'numeric' value. The minPoints is the minimum number
of points to construct a cluster.}

\item{\code{clustering}}{A 'matrix' that contains the result of one DBSCAN
clustering solution.}
}}

\section{Constructor}{


    Dbscan(name = "character", epsilon = "numeric", minPoints = "numeric",
        clustering = "matrix")

    name:       Empty character string or the name of the tSNE. \cr
    epsilon:    Empty 'numeric' representing the epsilon. \cr
    minPoints:  Empty 'numeric' representing the minPoints value. \cr
    clustering: Empty 'numeric' "matrix" or matrix of clustering. \cr
}

\section{Accessors}{


    In the following snippets, x is a Dbscan object.

    getName(x): Get the name of the Dbscan. \cr
    getEpsilon(x):    Get the epsilon used. \cr
    getMinPoints(x):  Get the MinPoint used. \cr
    getClustering(x): Get the matrix of DBSCAN clustering. \cr
}

\section{Subsetting}{


    In the following snippets, x is a Dbscan object.

    setName(x) <- value: Set the name of the Dbscan. \cr
    setEpsilon(x) <- value:    Set the epsilon used. \cr
    setMinPoints(x) <- value:  Set the minPoints used. \cr
    setClustering(x) <- value: Set the matrix of Dbscan clustering. \cr
}

\seealso{
runDBSCAN
}
\author{
Ilyess Rachedi and Nicolas Descostes
}
