% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getters}
\alias{getters}
\alias{getExperimentName,scRNAseq-method}
\alias{getExperimentName}
\alias{getCountMatrix,scRNAseq-method}
\alias{getCountMatrix}
\alias{getSceNorm,scRNAseq-method}
\alias{getSceNorm}
\alias{getSpecies,scRNAseq-method}
\alias{getSpecies}
\alias{getOutputDirectory,scRNAseq-method}
\alias{getOutputDirectory}
\alias{getTSNEList,scRNAseq-method}
\alias{getTSNEList}
\alias{getDbscanList,scRNAseq-method}
\alias{getDbscanList}
\alias{getCellsSimilarityMatrix,scRNAseq-method}
\alias{getCellsSimilarityMatrix}
\alias{getClustersSimilarityMatrix,scRNAseq-method}
\alias{getClustersSimilarityMatrix}
\alias{getClustersSimilarityOrdered,scRNAseq-method}
\alias{getClustersSimilarityOrdered}
\alias{getMarkerGenesList,scRNAseq-method}
\alias{getMarkerGenesList}
\alias{getClustersMarkers,scRNAseq-method}
\alias{getClustersMarkers}
\alias{getGenesInfos,scRNAseq-method}
\alias{getGenesInfos}
\alias{getName,Tsne-method}
\alias{getPerplexity,Tsne-method}
\alias{getPerplexity}
\alias{getPC,Tsne-method}
\alias{getPC}
\alias{getCoordinates,Tsne-method}
\alias{getCoordinates}
\alias{getName,Dbscan-method}
\alias{getName}
\alias{getEpsilon,Dbscan-method}
\alias{getEpsilon}
\alias{getMinPoints,Dbscan-method}
\alias{getMinPoints}
\alias{getClustering,Dbscan-method}
\alias{getClustering}
\title{getters}
\usage{
getExperimentName(theObject)

getCountMatrix(theObject)

getSceNorm(theObject)

getSpecies(theObject)

getOutputDirectory(theObject)

getTSNEList(theObject)

getDbscanList(theObject)

getCellsSimilarityMatrix(theObject)

getClustersSimilarityMatrix(theObject)

getClustersSimilarityOrdered(theObject)

getMarkerGenesList(theObject)

getClustersMarkers(theObject)

getGenesInfos(theObject)

\S4method{getName}{Tsne}(theObject)

getPerplexity(theObject)

getPC(theObject)

getCoordinates(theObject)

getName(theObject)

getEpsilon(theObject)

getMinPoints(theObject)

getClustering(theObject)
}
\arguments{
\item{theObject}{A scRNA-seq, Tsne or Dbscan object. See description or
?scRNAseq, ?Tsne, ?Dbscan.}
}
\value{
getExperimentName: Get the name of the experiment (scRNA-seq).

getCountMatrix: Get the count matrix (scRNA-seq).

getSceNorm: Get the SingleCellExperiment object used (scRNA-seq).

getSpecies: Get the species (scRNA-seq).

getOutputDirectory: Get the path of the output directory (scRNA-seq).

getTSNEList: Get the list of Tsne objects (scRNA-seq).

getDbscanList: Get the list of Dbscan objects (scRNA-seq).

getCellsSimilarityMatrix: Get the cell similarity matrix (scRNA-seq).

getClustersSimilarityMatrix: Get the cluster similarity matrix (scRNA-seq).

getClustersSimilarityOrdered: Get the clusters ordered by similarity
(scRNA-seq).

getMarkerGenesList: Get the list of marker genes by clusters  (scRNA-seq).

getClustersMarkers: Get the most significant markers by clusters (scRNA-seq).

getGenesInfos: Get informations about marker genes  (scRNA-seq).

getPerplexity: Get the perplexity used (Tsne).

getPC: Get the PC used (Tsne).

getCoordinates: Get the matrix of tSNE coordinates  (Tsne).

getName: Get the name of the tSNE or Dbscan object (Dbscan).

getEpsilon: Get the epsilon used (Dbscan).

getMinPoints: Get the MinPoint used (Dbscan).

getClustering: Get the matrix of DBSCAN clustering (Dbscan).
}
\description{
Retrieve the data of the slots of a scRNA-seq, Tsne or Dbscan object.
}
\examples{
## Load the count matrix
countmatrixPath <- system.file("extdata/example_countMatrix.tsv",
                            package="conclus")
countMatrix <- loadDataOrMatrix(file=countmatrixPath, type="countMatrix")

## Load the coldata
coldataPath <- system.file("extdata/example_colData.tsv",
                            package="conclus")
columnsMetaData <- loadDataOrMatrix(file=coldataPath, type="coldata",
columnID="cell_ID")

## Create the initial object
scr <- singlecellRNAseq(experimentName = "Bergiers",
                countMatrix     = countMatrix,
                species         = "mouse",
                outputDirectory = "YourOutputDirectory")

experimentName <- getExperimentName(scr)
countMatrix <- getCountMatrix(scr)
species <- getSpecies(scr)
outputDirectory <- getOutputDirectory(scr)

}
\author{
Ilyess RACHEDI
}
