\name{initialisePath}
\alias{initialisePath}
\title{Create all needed directories for CONCLUS output.}
\description{
Create all needed directories for CONCLUS output.
}
\usage{
initialisePath(dataDirectory)
}
\arguments{
\item{dataDirectory}{Output directory for a given CONCLUS run (supposed to be the same for one experiment during the workflow).}
}
