% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataset.R
\name{loadColdata}
\alias{loadColdata}
\title{loadColdata}
\usage{
loadColdata(file=coldataPath, columnCells="cellName", header=TRUE, 
            dec=".", sep='\t')
}
\arguments{
\item{file}{Path of the coldata to load.}

\item{columnCells}{Column name or indice of the column containing 
cell names/id.}

\item{header}{Set TRUE if the first row of the table corresponds to the
column names, and FALSE if it doesn't.}

\item{sep}{Character used in the table to separate the fields. 
Usually it's ' ' ';' ',' or '\t'.}

\item{dec}{Character used in the table for decimal points.
Usually '.' or ',' .}
}
\value{
The coldata with cellName column
}
\description{
This function import the coldata. All parameters are mandatory and no
default values, so the user has to know how is the coldata before the 
import. This function also replaces the name of the column containing cell
names and informed with 'columnCells' parameter by "cellName" for the good 
walk of CONCLUS.
}
\examples{
coldataPath <- file.path(system.file("extdata", package = "conclus"),
                        "test_colData_filtered.tsv")
                        
loadColdata(file=coldataPath, columnCells="cellName", header=TRUE, 
            dec=".", sep='\t')


}
