% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataset.R
\name{loadDataOrMatrix}
\alias{loadDataOrMatrix}
\title{loadDataOrMatrix}
\usage{
loadDataOrMatrix(file, type, columnID=NULL, header=TRUE, sep='\t', dec=".")
}
\arguments{
\item{file}{Path to the rowData, colData or Matrix.}

\item{type}{Values should be "coldata", "rowdata", or "countMatrix".}

\item{columnID}{For row and col data, column name containing cells/genes
names/id. Should not be used when inporting a matrix. Default=NULL.}

\item{header}{Set TRUE if the first row of the table corresponds to the
column names, and FALSE if it doesn't. Default=TRUE.}

\item{sep}{Character used in the table to separate the fields.
Usually it's ' ' ';' ',' or '\\t'. Default='\\t'.}

\item{dec}{Character used in the table for decimal points.
Usually '.' or ',' . Default=".".}
}
\value{
The formatted row, col data or the matrix.
}
\description{
This function allows to import the coldata, rowData or the countMatrix. It
formats each type of data to follow the requirements of CONCLUS.
}
\examples{

## ColData
coldataPath <- system.file("extdata/example_colData.tsv",
                            package="conclus")
loadDataOrMatrix(file=coldataPath, type="coldata", columnID="cell_ID")

## RowData
rowdataPath <- system.file("extdata/unit_tests_rowData_filtered.tsv",
                            package="conclus")
loadDataOrMatrix(file=rowdataPath, type="rowdata", columnID="gene_ID")

## Count matrix
countmatrixPath <- system.file("extdata/example_countMatrix.tsv",
                            package="conclus")
countMatrix <- loadDataOrMatrix(file=countmatrixPath, type="countMatrix")


}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES
}
