% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataset.R
\name{loadDataOrMatrix}
\alias{loadDataOrMatrix}
\title{loadDataOrMatrix}
\usage{
loadDataOrMatrix(file, type, columnID=NULL, header=TRUE, sep='\t', dec=".",
                    ignoreCellNumber=FALSE)
}
\arguments{
\item{file}{Path to the rowData, colData or Matrix.}

\item{type}{Values should be "coldata", "rowdata", or "countMatrix".}

\item{columnID}{For row and col data, column name containing cells/genes
names/id. Should not be used when inporting a matrix. Default=NULL.}

\item{header}{Set TRUE if the first row of the table corresponds to the
column names, and FALSE if it doesn't. Default=TRUE.}

\item{sep}{Character used in the table to separate the fields.
Usually it's ' ' ';' ',' or '\\t'. Default='\\t'.}

\item{dec}{Character used in the table for decimal points.
Usually '.' or ',' . Default=".".}

\item{ignoreCellNumber}{CONCLUS needs a large number of cells to collect
statistics, so we recommend using CONCLUS if you have at least 100 cells.
If you still want to use a count matrix with a number of cells inferior 
to 100 cells, set ignoreCellNumber=TRUE. Default=FALSE.}
}
\value{
The formatted row, col data or the matrix.
}
\description{
This function allows to import the coldata, rowData or the countMatrix. It
formats each type of data to follow the requirements of CONCLUS.
}
\examples{

## ColData
coldataPath <- system.file("extdata/colData.tsv", package="conclus")
loadDataOrMatrix(file=coldataPath, type="coldata", columnID="cell_ID")

## RowData
rowdataPath <- system.file("extdata/rowData.tsv",  package="conclus")
loadDataOrMatrix(file=rowdataPath, type="rowdata", columnID="gene_ID")

## Count matrix
countmatrixPath <- system.file("extdata/countMatrix.tsv", package="conclus")
countMatrix <- loadDataOrMatrix(file=countmatrixPath, type="countMatrix",
                                ignoreCellNumber=TRUE)


}
\author{
Ilyess RACHEDI and Nicolas DESCOSTES
}
