% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-normalization.R
\name{normaliseCountMatrix}
\alias{normaliseCountMatrix}
\alias{normaliseCountMatrix,normaliseCountMatrix-method}
\alias{normaliseCountMatrix,scRNAseq-method}
\alias{normalizeCountMatrix}
\title{normaliseCountMatrix() (Conclus class)}
\usage{
normaliseCountMatrix(
  theObject,
  sizes = c(20, 40, 60, 80, 100),
  rowdata = NULL,
  coldata = NULL,
  alreadyCellFiltered = FALSE,
  runQuickCluster = TRUE
)

\S4method{normaliseCountMatrix}{scRNAseq}(
  theObject,
  sizes = c(20, 40, 60, 80, 100),
  rowdata = NULL,
  coldata = NULL,
  alreadyCellFiltered = FALSE,
  runQuickCluster = TRUE
)
}
\arguments{
\item{theObject}{Object of class Conclus}

\item{sizes}{Vector of size factors from scran::computeSumFactors() function.}

\item{rowdata}{Optionnal. Data frame containing genes informations. Will be
create if it not provides by the user.}

\item{coldata}{Optionnal. Data frame containing cells informations. Will be
create if it not provides by the user.}

\item{alreadyCellFiltered}{Logical. If TRUE, quality check and
filtering will not be applied.}

\item{runQuickCluster}{Logical. If TRUE scran::quickCluster() function will
be applied. It usually improves the normalization for medium-size count
matrices. However, it is not recommended for datasets with less than 200
cells and may take too long for datasets with more than 10000 cells.}
}
\value{
Returns the Conclus object with the SCEnorm slot filled. It's an
SingleCellExperiment object containin the count matrix normalized, and also
the colData (table with cells informations) and the rowData (table with the
genes informations).
}
\description{
This function uses coldata (cells informations)
and rowdata (genes informations) to filters the count matrix. Then normalizes
it by using deconvolution with size factors method.
}
\author{
Ilyess RACHEDI, based on code by Polina PAVLOVICH
        and Nicolas DESCOSTES.
}
