% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_and_clustering_functions.R
\name{plotCellHeatmap}
\alias{plotCellHeatmap}
\title{Save markers heatmap.}
\usage{
plotCellHeatmap(markersClusters, sceObject, dataDirectory, experimentName,
  fileName, meanCentered = TRUE, colorPalette = "default",
  statePalette = "default", clusteringMethod = "ward.D2",
  orderClusters = FALSE, orderGenes = FALSE, returnPlot = FALSE,
  saveHeatmapTable = FALSE, width = 10, height = 8.5, ...)
}
\arguments{
\item{markersClusters}{a data frame where the first column is "geneName" containing genes names from sceObject, 
and the second column is corresponding "clusters". All names from that column must come from the column "clusters" in the colData(sceObject).
The data frame can be obtained from conclus::getMarkerGenes() function or created manually.}

\item{sceObject}{a SingleCellExperiment object with your experiment.}

\item{dataDirectory}{output directory of a given CONCLUS run (supposed to be the same for one experiment during the workflow).}

\item{experimentName}{name of the experiment which appears in filenames (supposed to be the same for one experiment during the workflow).}

\item{fileName}{name of the ouput file}

\item{meanCentered}{boolean, should mean centering be applied to the expression data or not.}

\item{colorPalette}{"default" or a vector of colors for the column "clusters" in the colData, for example c("yellow", "#CC79A7").}

\item{statePalette}{"default" or a vector of colors for the column "state" in the colData, for example c("yellow", "#CC79A7").}

\item{clusteringMethod}{a clustering methods passed to hclust() function.}

\item{orderClusters}{boolean, should the heatmap be structured by clusters.}

\item{orderGenes}{boolean, should the heatmap be structured by genes.}

\item{returnPlot}{boolean, whether to return a ggplot object with the plot or not.}

\item{saveHeatmapTable}{boolean, whether to save the expression matrix used for heatmap into a .csv file or not.
The file will be saved into 'dataDirectory/output_tables' with the same name as the .pdf plot.}

\item{width}{plot width.}

\item{height}{plot height.}

\item{...}{other parameters from pdf() and pheatmap() functions.}
}
\value{
A ggplot object of the plot if needed. The function saves pdf in "dataDirectiry/pictures" folder.
}
\description{
This function plots heatmap with marker genes on rows and clustered cells on columns.
}
