% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_and_clustering_functions.R
\name{plotClustersSimilarity}
\alias{plotClustersSimilarity}
\title{Save a similarity cluster matrix.}
\usage{
plotClustersSimilarity(clustersSimilarityMatrix, sceObject, dataDirectory,
  experimentName, colorPalette, statePalette, clusteringMethod,
  returnPlot = FALSE, width = 7, height = 5.5, ...)
}
\arguments{
\item{clustersSimilarityMatrix}{a matrix, result of conclus::calculateClustersSimilarity() function.}

\item{sceObject}{a SingleCellExperiment object with your experiment.}

\item{dataDirectory}{output directory for CONCLUS (supposed to be the same for one experiment during the workflow).}

\item{experimentName}{name of the experiment which will appear in filenames (supposed to be the same for one experiment during the workflow).}

\item{colorPalette}{"default" or a vector of colors for the column "clusters" in the colData, for example c("yellow", "#CC79A7").}

\item{statePalette}{"default" or a vector of colors for the column "state" in the colData, for example c("yellow", "#CC79A7").}

\item{clusteringMethod}{a clustering methods passed to hclust() function.}

\item{returnPlot}{boolean, return plot or not.}

\item{width}{plot width.}

\item{height}{plot height.}

\item{...}{other parameters of pdf() and pheatmap() functions.}
}
\value{
A ggplot object or nothing (depends on returnPlot parameter). It saves the pdf in "dataDirectory/pictures" folder.
}
\description{
Save a similarity cluster matrix.
}
