% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{plotGeneExpression,scRNAseq-method}
\alias{plotGeneExpression,scRNAseq-method}
\alias{plotGeneExpression}
\title{plotGeneExpression}
\usage{
plotGeneExpression(theObject, geneName,
palette=c("grey","red", "#7a0f09", "black"), returnPlot=FALSE,
            tSNEpicture=1, savePlot=FALSE, alpha=1, limits=NA,
            pointSize=1, width=6, height=5, plotPDF=TRUE, silentPlot=FALSE)
}
\arguments{
\item{theObject}{A scRNAseq object with the top markers retrieved. See
?retrieveTopClustersMarkers.}

\item{geneName}{Name of the gene to highlight on the t-SNE plot.}

\item{palette}{Color palette for the expression levels.}

\item{returnPlot}{If TRUE, returns a ggplot object of the tSNE.
Default = FALSE.}

\item{tSNEpicture}{Number of the tSNE picture that you want to use for
plotting the gene expression. Default = 1.}

\item{savePlot}{If TRUE, save the tSNE in pdf or png format. Default=FALSE.}

\item{alpha}{Opacity of the points of the plot. Default = 1.}

\item{limits}{Range of the gene expression shown in the legend. Default = NA.
See details.}

\item{pointSize}{Size of the points on the tSNE. Default = 1.}

\item{width}{Width of the plot. Default = 6.}

\item{height}{Height of the plot. Default = 5.}

\item{plotPDF}{If TRUE export tSNE in pdf format, if FALSE export it in
png format. Default=TRUE.}

\item{silentPlot}{If TRUE, the plots are not displayed on the current device.
Default=FALSE.}
}
\value{
A ggplot object of the gene expression colored tSNE.
}
\description{
The function saves a t-SNE plot colored by expression of a
given gene.
}
\details{
limits -- This option allows generating t-SNE plots with equal color scale to
compare the expression of different genes. By default, limits are the range
of expression of a selected gene.
}
\examples{
## Load the count matrix
countmatrixPath <- system.file("extdata/test_countMatrix.tsv", 
                            package="conclus")
countMatrix <- loadDataOrMatrix(file=countmatrixPath, type="countMatrix")

## Load the coldata
coldataPath <- system.file("extdata/test_colData_filtered.tsv", 
                            package="conclus")
columnsMetaData <- loadDataOrMatrix(file=coldataPath, type="coldata", 
columnID="cell_ID")

## Create the initial object
scr <- singlecellRNAseq(experimentName = "Bergiers",
                countMatrix     = countMatrix,
                species         = "mouse",
                outputDirectory = "YourOutputDirectory")

## Normalize and filter the raw counts matrix
scrNorm <- normaliseCountMatrix(scr, coldata = columnsMetaData)

## Compute the tSNE coordinates
scrTsne <- generateTSNECoordinates(scrNorm, cores=2)

## Perform the clustering with dbScan
scrDbscan <- runDBSCAN(scrTsne, cores=2)

## Compute the cell similarity matrix
scrCCI <- clusterCellsInternal(scrDbscan, clusterNumber=10, cores=2)

## Calculate clusters similarity
scrCSM <- calculateClustersSimilarity(scrCCI)

## Ranking genes
scrS4MG <- rankGenes(scrCSM)

## Retrieve top clusters markers
scrFinal <- retrieveTopClustersMarkers(scrS4MG, removeDuplicates=FALSE)

## t-SNE plot colored by expression of a  given gene.
plotGeneExpression(scrFinal, getClustersMarkers(scrFinal)[1,1])

}
\seealso{
retrieveTopClustersMarkers plotCellSimilarity plotCellHeatmap
plotClusteredTSNE plotClustersSimilarity
}
\author{
Ilyess RACHEDI, based on code by Polina PAVLOVICH and Nicolas DESCOSTES.
}
