% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-clustering.R
\name{retrieveTableClustersCells,scRNAseq-method}
\alias{retrieveTableClustersCells,scRNAseq-method}
\alias{retrieveTableClustersCells}
\title{retrieveTableClustersCells}
\usage{
retrieveTableClustersCells(theObject)
}
\arguments{
\item{theObject}{An Object of class scRNASeq for which the cells were
clustered internally. See ?clusterCellsInternal.}
}
\value{
A data frame containing two columns 'clusters' and 'cells' indicating the
result of the consensus clustering at the cellular level.
}
\description{
Having computed clusterCellsInternal, retrieve to what cluster each cell
belongs. The output data.frame can be passed to the method ?addClustering.
}
\examples{
## Object scr containing the results of previous steps
load(system.file("extdata/scrFull.Rdat", package="conclus"))

## Retrieving the table clusters-cells.
cellClustDf <- retrieveTableClustersCells(scr)

}
\seealso{
addClustering
}
\author{
Nicolas DESCOSTES.
}
