% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-markers.R
\name{retrieveTopClustersMarkers,scRNAseq-method}
\alias{retrieveTopClustersMarkers,scRNAseq-method}
\alias{retrieveTopClustersMarkers}
\title{retrieveTopClustersMarkers}
\usage{
retrieveTopClustersMarkers(theObject, nTop=10, removeDuplicates = TRUE,
				writeMarkerGenes = FALSE)
}
\arguments{
\item{theObject}{An Object of class scRNASeq for which rankGenes was 
run. See ?rankGenes.}

\item{nTop}{Number of marker genes to retrieve per cluster. Default=10.}

\item{removeDuplicates}{If TRUE, duplicated markers are removed from the 
lists. Default=TRUE.}

\item{writeMarkerGenes}{If TRUE, writes one list per cluster in the output
folder defined in theObject, and in the sub-directory 
marker_genes/markers_lists. Default=FALSE.}
}
\value{
Output the list of markers to marker_genes/markers_lists if writeMarkersGenes
is TRUE and return a scRNASeq object with its clustersMarkers slot updated.
}
\description{
This function retrieves the top N marker genes for each cluster.
}
\examples{
experimentName <- "Bergiers"
countMatrix <- as.matrix(read.delim(file.path(
"tests/testthat/test_data/test_countMatrix.tsv")))
outputDirectory <- "./"
columnsMetaData <- read.delim(
file.path("extdata/Bergiers_colData_filtered.tsv"))

## Create the initial object
scr <- scRNAseq(experimentName = experimentName, 
                countMatrix     = countMatrix, 
                species         = "mouse",
                outputDirectory = outputDirectory)

## Normalize and filter the raw counts matrix
scrNorm <- normaliseCountMatrix(scr, coldata = columnsMetaData)

## Compute the tSNE coordinates
scrTsne <- generateTSNECoordinates(scrNorm, cores=5)

## Perform the clustering with dbScan
scrDbscan <- runDBSCAN(scrTsne, cores=5)

## Compute the cell similarity matrix
scrCCI <- clusterCellsInternal(scrDbscan, clusterNumber=10, cores=4)

## Calculate clusters similarity
scrCSM <- calculateClustersSimilarity(scrCCI)

## Ranking genes
scrS4MG <- rankGenes(scrCSM)

## Retrieve the top 10 markers per cluster
scrFinal <- retrieveTopClustersMarkers(theObject)

}
\seealso{
retrieveGenesInfo
}
\author{
Ilyess RACHEDI, based on code by Polina PAVLOVICH and Nicolas DESCOSTES.
}
