% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-markers.R
\name{retrieveTopClustersMarkers,scRNAseq-method}
\alias{retrieveTopClustersMarkers,scRNAseq-method}
\alias{retrieveTopClustersMarkers}
\title{retrieveTopClustersMarkers}
\usage{
retrieveTopClustersMarkers(theObject, nTop=10, removeDuplicates = TRUE,
                writeMarkerGenes = FALSE)
}
\arguments{
\item{theObject}{An Object of class scRNASeq for which rankGenes was
run. See ?rankGenes.}

\item{nTop}{Number of marker genes to retrieve per cluster. Default=10.}

\item{removeDuplicates}{If TRUE, duplicated markers are removed from the
lists. Default=TRUE.}

\item{writeMarkerGenes}{If TRUE, writes one list per cluster in the output
folder defined in theObject, and in the sub-directory
marker_genes/markers_lists. Default=FALSE.}
}
\value{
Output the list of markers to marker_genes/markers_lists if writeMarkersGenes
is TRUE and return a scRNASeq object with its clustersMarkers slot updated.
}
\description{
This function retrieves the top N marker genes for each cluster.
}
\examples{
## Object scr containing the results of previous steps
load(system.file("extdata/scrFull.Rdat", package="conclus"))

## Retrieve the top 10 markers per cluster
scr <- retrieveTopClustersMarkers(scr)

}
\seealso{
retrieveGenesInfo
}
\author{
Ilyess RACHEDI, based on code by Polina PAVLOVICH and Nicolas DESCOSTES.
}
