% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_and_clustering_functions.R
\name{runCONCLUS}
\alias{runCONCLUS}
\title{Run CONCLUS in one click}
\usage{
runCONCLUS(sceObject, dataDirectory, experimentName,
  colorPalette = "default", statePalette = "default",
  clusteringMethod = "ward.D2", epsilon = c(1.3, 1.4, 1.5),
  minPoints = c(3, 4), k = 0, PCs = c(4, 6, 8, 10, 20, 40, 50),
  perplexities = c(30, 40), randomSeed = 42, deepSplit = 4,
  preClustered = F, orderClusters = FALSE, cores = 14,
  plotPDFcellSim = TRUE, deleteOutliers = TRUE,
  tSNEalreadyGenerated = FALSE, tSNEresExp = "")
}
\arguments{
\item{sceObject}{a SingleCellExperiment object with your data.}

\item{dataDirectory}{CONCLUS will create this directory if it doesn't exist and store there all output files.}

\item{experimentName}{most of output file names of CONCLUS are hardcoded.
experimentName will stay at the beginning of each output file name to
distinguish different runs easily.}

\item{colorPalette}{a vector of colors for clusters.}

\item{statePalette}{a vector of colors for states.}

\item{clusteringMethod}{a clustering methods passed to hclust() function.}

\item{epsilon}{a parameter of fpc::dbscan() function.}

\item{minPoints}{a parameter of fpc::dbscan() function.}

\item{k}{preferred number of clusters. Alternative to deepSplit. A parameter of cutree() function.}

\item{PCs}{a vector of first principal components.
For example, to take ranges 1:5 and 1:10 write c(5, 10).}

\item{perplexities}{a vector of perplexity for t-SNE.}

\item{randomSeed}{random seed for reproducibility.}

\item{deepSplit}{intuitive level of clustering depth. Options are 1, 2, 3, 4.}

\item{preClustered}{if TRUE, it will not change the column clusters after the run.
However, it will anyway run DBSCAN to calculate similarity matrices.}

\item{orderClusters}{can be either FALSE (default) of "name".
If "name", clusters in the similarity matrix of cells will be ordered by name.}

\item{cores}{maximum number of jobs that CONCLUS can run in parallel.}

\item{plotPDFcellSim}{if FALSE, the similarity matrix of cells will be saved in png format.
FALSE is recommended for count matrices with more than 2500 cells due to large pdf file size.}

\item{deleteOutliers}{whether cells which were often defined as outliers by dbscan must be deleted.
It will require recalculating of the similarity matrix of cells. Default is FALSE.
Usually those cells form a separate "outlier" cluster and can be easier distinguished and deleted later
if necessary.}

\item{tSNEalreadyGenerated}{if you already ran CONCLUS ones and have t-SNE coordinated saved
You can set TRUE to run the function faster since it will skip the generation of t-SNE coordinates and use the stored ones. 
Option TRUE requires t-SNE coordinates to be located in your 'dataDirectory/tsnes' directory.}

\item{tSNEresExp}{experimentName of t-SNE coordinates which you want to use.
This argument allows copying and pasting t-SNE coordinates between different CONCLUS runs without renaming the files.}
}
\value{
A SingleCellExperiment object.
}
\description{
This function performs core CONCLUS workflow. It generates PCA and t-SNE coordinates, 
runs DBSCAN, calculates similarity matrices of cells and clusters, assigns cells to clusters,
searches for positive markers for each cluster. The function saves plots and tables into dataDirectory.
}
\keyword{CONCLUS}
