% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{scRNAseq-class}
\alias{scRNAseq-class}
\alias{scRNAseq}
\title{The scRNAseq class}
\description{
S4 class and the main class used by CONCLUS containing the results of
the different steps to analyse rare cell populations.
}
\section{Slots}{

\describe{
\item{\code{experimentName}}{'character' string representing the name of the
experiment.}

\item{\code{countMatrix}}{An 'integer matrix' representing the raw count matrix
with reads or unique molecular identifiers (UMIs).}

\item{\code{sceNorm}}{Object of class SingleCellExperiment that contains the
colData giving informations about cells and the rowData giving
informations about genes. It also contains the normalized count matrix.}

\item{\code{species}}{'character' string representing the species of interest.
Currently limited to "mouse" and "human". Other organisms can be 
added on demand.}

\item{\code{outputDirectory}}{A 'character' string of the path to the root
output folder.}

\item{\code{tSNEList}}{List of 'Tsne' objects representing the different tSNE
coordinates generated by CONCLUS.}

\item{\code{dbscanList}}{List of 'Dbscan' objects representing the different 
Dbscan clustering generated by CONCLUS.}

\item{\code{cellsSimilarityMatrix}}{A numeric Matrix defining how many times
two cells have been associated to the same cluster across the 84
solutions (by default) of clustering.}

\item{\code{clustersSimilarityMatrix}}{A numeric matrix comparing the
robustness of the consensus clusters.}

\item{\code{clustersSimiliratyOrdered}}{A factor representing the clusters 
ordered by similarity.}

\item{\code{markerGenesList}}{List of data.frames. Each data frame contains
the ranked genes of one cluster.}

\item{\code{clustersMarkers}}{A data frame containing the top 10 (by default)
marker genes of each clusters.}

\item{\code{genesInfos}}{A data frame containing informations of the markers
genes for each clusters.}
}}

\section{Constructor}{


singlecellRNAseq(experimentName = "character", countMatrix = "matrix",
species = "character", outputDirectory = "character")

experimentName:  String of the name of the experiment.

countMatrix:     Matrix containing the raw counts.

species:         'character' string representing the species of interest. 
                Shoud be mouse or human. Other organisms can be added on 
                demand.

outputDirectory: 'character' string representing the path to the output 
directory.
}

\section{Accessors}{


    In the following snippets, x is a scRNAseq object.

    getExperimentName(x):            Get the name of the experiment. \cr
    getCountMatrix(x):               Get the count matrix. \cr
    getSceNorm(x):                   Get the SingleCellExperiment object 
used  \cr
    getSpecies(x):                   Get the species. \cr
    getOutputDirectory(x):           Get the path of the output directory. 
\cr
    getTSNEList(x):                  Get the list of Tsne objects. \cr
    getDbscanList(x):                Get the list of Dbscan objects. \cr
    getCellsSimilarityMatrix(x):     Get the cell similarity matrix. \cr
    getClustersSimilarityMatrix(x):  Get the cluster similarity matrix. \cr
    getClustersSimiliratyOrdered(x): Get the clusters ordered by 
similarity. \cr
    getMarkerGenesList(x):           Get the list of marker genes by 
clusters. \cr
    getClustersMarkers(x):           Get the most significant markers by
                                    clusters into a data.frame. \cr
    getGenesInfos(x):                Get a data frame containing informations
                                    about marker genes. \cr
}

\section{Subsetting}{


    In the following snippets, x is a scRNAseq object.

    setExperimentName(x):            Set the name of the experiment. \cr
    setCountMatrix(x):               Set the count matrix. \cr
    setSceNorm(x):                   Set the SingleCellExperiment object 
used. \cr
    setSpecies(x):                   Set the species. \cr
    setOutputDirectory(x):           Set the path of the output directory. 
\cr
    setTSNEList(x):                  Set the list of Tsne objects. \cr
    setDbscanList(x):                Set the list of Dbscan objects. \cr
    setCellsSimilarityMatrix(x):     Set the cell similarity matrix. \cr
    setClustersSimilarityMatrix(x):  Set the cluster similarity matrix. \cr
    setClustersSimiliratyOrdered(x): Set the clusters ordered by 
similarity. \cr
    setMarkerGenesList(x):           Set the list of marker genes by 
clusters \cr
    setClustersMarkers(x):           Set the most significant markers by
                                    clusters. \cr
    setGenesInfos(x):                Set a data.frame containing informations
                                    about the marker genes. \cr
}

\seealso{
singlecellRNAseq
}
\author{
Ilyess Rachedi and Nicolas Descostes
}
