% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/getters.R, R/setters.R
\docType{class}
\name{scRNAseq-class}
\alias{scRNAseq-class}
\alias{scRNAseq}
\alias{getExperimentName}
\alias{getCountMatrix}
\alias{getSceNorm}
\alias{getSpecies}
\alias{getOutputDirectory}
\alias{getTSNEList}
\alias{getDbscanList}
\alias{getCellsSimilarityMatrix}
\alias{getClustersSimilarityMatrix}
\alias{getClustersSimiliratyOrdered}
\alias{getMarkerGenesList}
\alias{getClustersMarkers}
\alias{getGenesInfos}
\alias{setExperimentName}
\alias{setCountMatrix}
\alias{setSceNorm}
\alias{setSpecies}
\alias{setOutputDirectory}
\alias{setTSNEList}
\alias{setDbscanList}
\alias{setCellsSimilarityMatrix}
\alias{setClustersSimilarityMatrix}
\alias{setClustersSimiliratyOrdered}
\alias{setMarkerGenesList}
\alias{setClustersMarkers}
\alias{setGenesInfos}
\alias{getExperimentName,scRNAseq-method}
\alias{getCountMatrix,scRNAseq-method}
\alias{getSceNorm,scRNAseq-method}
\alias{getSpecies,scRNAseq-method}
\alias{getOutputDirectory,scRNAseq-method}
\alias{getTSNEList,scRNAseq-method}
\alias{getDbscanList,scRNAseq-method}
\alias{getCellsSimilarityMatrix,scRNAseq-method}
\alias{getClustersSimilarityMatrix,scRNAseq-method}
\alias{getClustersSimiliratyOrdered,scRNAseq-method}
\alias{getMarkerGenesList,scRNAseq-method}
\alias{getClustersMarkers,scRNAseq-method}
\alias{getGenesInfos,scRNAseq-method}
\alias{setExperimentName<-,scRNAseq-method}
\alias{setCountMatrix<-,scRNAseq-method}
\alias{setSceNorm<-,scRNAseq-method}
\alias{setSpecies<-,scRNAseq-method}
\alias{setOutputDirectory<-,scRNAseq-method}
\alias{setTSNEList<-,scRNAseq-method}
\alias{setDbscanList<-,scRNAseq-method}
\alias{setCellsSimilarityMatrix<-,scRNAseq-method}
\alias{setClustersSimilarityMatrix<-,scRNAseq-method}
\alias{setClustersSimiliratyOrdered<-,scRNAseq-method}
\alias{setMarkerGenesList<-,scRNAseq-method}
\alias{setClustersMarkers<-,scRNAseq-method}
\alias{setGenesInfos<-,scRNAseq-method}
\title{The scRNAseq class}
\usage{
\S4method{getExperimentName}{scRNAseq}(theObject)

\S4method{getCountMatrix}{scRNAseq}(theObject)

\S4method{getSceNorm}{scRNAseq}(theObject)

\S4method{getSpecies}{scRNAseq}(theObject)

\S4method{getOutputDirectory}{scRNAseq}(theObject)

\S4method{getTSNEList}{scRNAseq}(theObject)

\S4method{getDbscanList}{scRNAseq}(theObject)

\S4method{getCellsSimilarityMatrix}{scRNAseq}(theObject)

\S4method{getClustersSimilarityMatrix}{scRNAseq}(theObject)

\S4method{getClustersSimiliratyOrdered}{scRNAseq}(theObject)

\S4method{getMarkerGenesList}{scRNAseq}(theObject)

\S4method{getClustersMarkers}{scRNAseq}(theObject)

\S4method{getGenesInfos}{scRNAseq}(theObject)

\S4method{setExperimentName}{scRNAseq}(theObject) <- value

\S4method{setCountMatrix}{scRNAseq}(theObject) <- value

\S4method{setSceNorm}{scRNAseq}(theObject) <- value

\S4method{setSpecies}{scRNAseq}(theObject) <- value

\S4method{setOutputDirectory}{scRNAseq}(theObject) <- value

\S4method{setTSNEList}{scRNAseq}(theObject) <- value

\S4method{setDbscanList}{scRNAseq}(theObject) <- value

\S4method{setCellsSimilarityMatrix}{scRNAseq}(theObject) <- value

\S4method{setClustersSimilarityMatrix}{scRNAseq}(theObject) <- value

\S4method{setClustersSimiliratyOrdered}{scRNAseq}(theObject) <- value

\S4method{setMarkerGenesList}{scRNAseq}(theObject) <- value

\S4method{setClustersMarkers}{scRNAseq}(theObject) <- value

\S4method{setGenesInfos}{scRNAseq}(theObject) <- value
}
\description{
S4 class and the main class used by CONCLUS containing the results of the 
different steps to analyse rare cell populations.
}
\section{Functions}{
\itemize{
\item \code{getExperimentName,scRNAseq-method}: Get the name of the experiment.

\item \code{getCountMatrix,scRNAseq-method}: Get the count matrix.

\item \code{getSceNorm,scRNAseq-method}: Get the SingleCellExperiment object used.

\item \code{getSpecies,scRNAseq-method}: Get the species.

\item \code{getOutputDirectory,scRNAseq-method}: Get the path of the output directory.

\item \code{getTSNEList,scRNAseq-method}: Get the list of Tsne objects.

\item \code{getDbscanList,scRNAseq-method}: Get the list of Dbscan objects.

\item \code{getCellsSimilarityMatrix,scRNAseq-method}: Get the cell similarity matrix.

\item \code{getClustersSimilarityMatrix,scRNAseq-method}: Get the cluster similarity matrix.

\item \code{getClustersSimiliratyOrdered,scRNAseq-method}: Get the clusters ordered by similarity.

\item \code{getMarkerGenesList,scRNAseq-method}: Get the list of marker genes by clusters

\item \code{getClustersMarkers,scRNAseq-method}: Get the most significant markers by clusters.

\item \code{getGenesInfos,scRNAseq-method}: Get informations about marker genes.

\item \code{setExperimentName<-,scRNAseq-method}: Set the name of the experiment.

\item \code{setCountMatrix<-,scRNAseq-method}: Set the count matrix.

\item \code{setSceNorm<-,scRNAseq-method}: Set the SingleCellExperiment object used.

\item \code{setSpecies<-,scRNAseq-method}: Set the species.

\item \code{setOutputDirectory<-,scRNAseq-method}: Set the path of the output directory.

\item \code{setTSNEList<-,scRNAseq-method}: Set the list of Tsne objects.

\item \code{setDbscanList<-,scRNAseq-method}: Set the list of Dbscan objects.

\item \code{setCellsSimilarityMatrix<-,scRNAseq-method}: Set the cell similarity matrix.

\item \code{setClustersSimilarityMatrix<-,scRNAseq-method}: Set the cluster similarity matrix.

\item \code{setClustersSimiliratyOrdered<-,scRNAseq-method}: Set the clusters ordered by similarity.

\item \code{setMarkerGenesList<-,scRNAseq-method}: Set the list of marker genes by clusters.

\item \code{setClustersMarkers<-,scRNAseq-method}: Set the most significant markers by clusters.

\item \code{setGenesInfos<-,scRNAseq-method}: Set a data.frame containing informations about 
the marker genes.
}}

\section{Slots}{

\describe{
\item{\code{experimentName}}{'character' string representing the name of the 
experiment. Many output of scRNAseq will use this name.}

\item{\code{countMatrix}}{An 'integer matrix' representing the raw count matrix with 
reads or unique molecular identifiers (UMIs).}

\item{\code{sceNorm}}{Object of class SingleCellExperiment that contains the colData 
giving informations about cells and the rowData giving informations about 
genes. It also contains the normalized count matrix.}

\item{\code{species}}{'character' string representing the species of interest. 
Currently limited to "mouse" and "human".}

\item{\code{outputDirectory}}{A 'character' string of the path to the root output 
folder.}

\item{\code{tSNEList}}{List of 'Tsne' objects representing the different tSNE 
coordinates generated by CONCLUS.}

\item{\code{dbscanList}}{List of 'Dbscan' objects representing the different Dbscan 
clustering generated by CONCLUS.}

\item{\code{cellsSimilarityMatrix}}{A numeric Matrix defining how many times two 
cells have been associated to the same cluster across the 84 solutions of 
clustering.}

\item{\code{clustersSimilarityMatrix}}{A numeric matrix comparing the robustness of 
the consensus clusters.}

\item{\code{clustersSimiliratyOrdered}}{A factor representing the clusters ordered by
similarity.}

\item{\code{markerGenesList}}{list of data.frames. Each data frame contains the 
ranked genes of one cluster.}

\item{\code{clustersMarkers}}{A data frame containing the top 10 (by default) marker 
genes of each clusters.}

\item{\code{genesInfos}}{A data frame containing informations of the markers genes 
for each clusters.}
}}

\section{Constructor}{


scRNAseq(experimentName = "character", countMatrix = "matrix", 
sceNorm = "SingleCellExperiment", species = "character", 
outputDirectory = "character", tSNEList = "list", dbscanList = "list",
cellsSimilarityMatrix = "matrix", clustersSimilarityMatrix = "matrix",
clustersSimiliratyOrdered = "factor", markerGenesList = "list", 
clustersMarkers = "data.frame", genesInfos = "data.frame")

experimentName:            String of the name of the experiment.
 
countMatrix:               Matrix containing the raw counts.

sceNorm:                   Empty or SingleCellExperiment object containing 
                           the normalized count matrix, the colData and the 
                           rowData.
species:                   'character' representing the species of interest. 
                           Shoud be mouse or human.

outputDirectory:           'character' representing the path to the output 
                           directory.

tSNEList:                  Empty or list of Tsne objects.

dbscanList:                Empty or list of Dbscan objects.

cellsSimilarityMatrix:     Empty or cells matrix with similarity scores.

clustersSimilarityMatrix:  Empty or clusters matrix with similarity scores.

clustersSimiliratyOrdered: Empty or factor representing the clusters ordered 
                           by similarity.

markerGenesList:           Empty or list of data frames. Each data frame 
                           contains the ranked genes of one cluster.

clustersMarkers:           Empty or data frame containing the top 10 (by 
                           default) marker genes of each clusters.

genesInfos:                Empty or dataframe containing informations about 
                           the markers genes each clusters.
}

\section{Accessors}{


    In the following snippets, x is a scRNAseq object.

    getExperimentName(x):            Get the name of the experiment.
    getCountMatrix(x):               Get the count matrix. 
    getSceNorm(x):                   Get the SingleCellExperiment object used
    getSpecies(x):                   Get the species.  
    getOutputDirectory(x):           Get the path of the output directory. 
    getTSNEList(x):                  Get the list of Tsne objects.
    getDbscanList(x):                Get the list of Dbscan objects.
    getCellsSimilarityMatrix(x):     Get the cell similarity matrix. 
    getClustersSimilarityMatrix(x):  Get the cluster similarity matrix.
    getClustersSimiliratyOrdered(x): Get the clusters ordered by similarity.
    getMarkerGenesList(x):           Get the list of marker genes by clusters
    getClustersMarkers(x):           Get the most significant markers by 
                                     clusters into a data.frame.
    getGenesInfos(x):                Get a data frame containing informations
                                     about marker genes.
}

\section{Subsetting}{


    In the following snippets, x is a scRNAseq object.

    setExperimentName(x):            Set the name of the experiment.
    setCountMatrix(x):               Set the count matrix. 
    setSceNorm(x):                   Set the SingleCellExperiment object used
    setSpecies(x):                   Set the species. 
    setOutputDirectory(x):           Set the path of the output directory. 
    setTSNEList(x):                  Set the list of Tsne objects.
    setDbscanList(x):                Set the list of Dbscan objects.
    setCellsSimilarityMatrix(x):     Set the cell similarity matrix. 
    setClustersSimilarityMatrix(x):  Set the cluster similarity matrix.
    setClustersSimiliratyOrdered(x): Set the clusters ordered by similarity.   
    setMarkerGenesList(x):           Set the list of marker genes by clusters
    setClustersMarkers(x):           Set the most significant markers by 
                                     clusters.
    setGenesInfos(x):                Set a data.frame containing informations
                                     about the marker genes.
}

\author{
Ilyess Rachedi and Nicolas Descostes
}
