% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{scRNAseq-class}
\alias{scRNAseq-class}
\alias{scRNAseq}
\title{The scRNAseq class}
\description{
S4 class and the main class used by CONCLUS containing the results of the 
different steps to analyse rare cell populations.
}
\section{Slots}{

\describe{
\item{\code{experimentName}}{'character' string representing the name of the 
experiment. Many output of scRNAseq will use this name.}

\item{\code{countMatrix}}{An 'integer matrix' representing the raw count matrix with 
reads or unique molecular identifiers (UMIs).}

\item{\code{sceNorm}}{Object of class SingleCellExperiment that contains the colData 
giving informations about cells and the rowData giving informations about 
genes. It also contains the normalized count matrix.}

\item{\code{species}}{'character' string representing the species of interest. 
Currently limited to "mouse" and "human".}

\item{\code{outputDirectory}}{A 'character' string of the path to the root output 
folder.}

\item{\code{tSNEList}}{List of 'Tsne' objects representing the different tSNE 
coordinates generated by CONCLUS.}

\item{\code{dbscanList}}{List of 'Dbscan' objects representing the different Dbscan 
clustering generated by CONCLUS.}

\item{\code{cellsSimilarityMatrix}}{A numeric Matrix defining how many times two 
cells have been associated to the same cluster across the 84 solutions of 
clustering.}

\item{\code{clustersSimilarityMatrix}}{A numeric matrix comparing the robustness of 
the consensus clusters.}

\item{\code{clustersSimiliratyOrdered}}{A factor representing the clusters ordered by
similarity.}

\item{\code{markerGenesList}}{list of data.frames. Each data frame contains the 
ranked genes of one cluster.}

\item{\code{clustersMarkers}}{A data frame containing the top 10 (by default) marker 
genes of each clusters.}

\item{\code{genesInfos}}{A data frame containing informations of the markers genes 
for each clusters.}
}}

\section{Constructor}{


scRNAseq(experimentName = "character", countMatrix = "matrix", 
sceNorm = "SingleCellExperiment", species = "character", 
outputDirectory = "character", tSNEList = "list", dbscanList = "list",
cellsSimilarityMatrix = "matrix", clustersSimilarityMatrix = "matrix",
clustersSimiliratyOrdered = "factor", markerGenesList = "list", 
clustersMarkers = "data.frame", genesInfos = "data.frame")

experimentName:            String of the name of the experiment.
 
countMatrix:               Matrix containing the raw counts.

sceNorm:                   Empty or SingleCellExperiment object containing 
                           the normalized count matrix, the colData and the 
                           rowData.
species:                   'character' representing the species of interest. 
                           Shoud be mouse or human.

outputDirectory:           'character' representing the path to the output 
                           directory.

tSNEList:                  Empty or list of Tsne objects.

dbscanList:                Empty or list of Dbscan objects.

cellsSimilarityMatrix:     Empty or cells matrix with similarity scores.

clustersSimilarityMatrix:  Empty or clusters matrix with similarity scores.

clustersSimiliratyOrdered: Empty or factor representing the clusters ordered 
                           by similarity.

markerGenesList:           Empty or list of data frames. Each data frame 
                           contains the ranked genes of one cluster.

clustersMarkers:           Empty or data frame containing the top 10 (by 
                           default) marker genes of each clusters.

genesInfos:                Empty or dataframe containing informations about 
                           the markers genes each clusters.
}

\section{Accessors}{


    In the following snippets, x is a scRNAseq object.

    getExperimentName(x):            Get the name of the experiment.
    getCountMatrix(x):               Get the count matrix. 
    getSceNorm(x):                   Get the SingleCellExperiment object used
    getSpecies(x):                   Get the species.  
    getOutputDirectory(x):           Get the path of the output directory. 
    getTSNEList(x):                  Get the list of Tsne objects.
    getDbscanList(x):                Get the list of Dbscan objects.
    getCellsSimilarityMatrix(x):     Get the cell similarity matrix. 
    getClustersSimilarityMatrix(x):  Get the cluster similarity matrix.
    getClustersSimiliratyOrdered(x): Get the clusters ordered by similarity.
    getMarkerGenesList(x):           Get the list of marker genes by clusters
    getClustersMarkers(x):           Get the most significant markers by 
                                     clusters into a data.frame.
    getGenesInfos(x):                Get a data frame containing informations
                                     about marker genes.
}

\section{Subsetting}{


    In the following snippets, x is a scRNAseq object.

    setExperimentName(x):            Set the name of the experiment.
    setCountMatrix(x):               Set the count matrix. 
    setSceNorm(x):                   Set the SingleCellExperiment object used
    setSpecies(x):                   Set the species. 
    setOutputDirectory(x):           Set the path of the output directory. 
    setTSNEList(x):                  Set the list of Tsne objects.
    setDbscanList(x):                Set the list of Dbscan objects.
    setCellsSimilarityMatrix(x):     Set the cell similarity matrix. 
    setClustersSimilarityMatrix(x):  Set the cluster similarity matrix.
    setClustersSimiliratyOrdered(x): Set the clusters ordered by similarity.   
    setMarkerGenesList(x):           Set the list of marker genes by clusters
    setClustersMarkers(x):           Set the most significant markers by 
                                     clusters.
    setGenesInfos(x):                Set a data.frame containing informations
                                     about the marker genes.
}

\author{
Ilyess Rachedi and Nicolas Descostes
}
