% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{setExperimentName<-,scRNAseq-method}
\alias{setExperimentName<-,scRNAseq-method}
\alias{setExperimentName}
\alias{setCountMatrix<-,scRNAseq-method}
\title{setExperimentName}
\usage{
setExperimentName(theObject) <- value

setCountMatrix(theObject) <- value
}
\arguments{
\item{theObject}{A scRNA-seq object to update.}

\item{value}{The value to update the slot with.}
}
\description{
setExperimentName

setCountMatrix
}
\section{Functions}{
\itemize{
\item \code{setCountMatrix<-,scRNAseq-method}: Set the count matrix.
}}

