% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\encoding{UTF-8}
\name{setters}
\alias{setters}
\alias{setExperimentName<-,scRNAseq-method}
\alias{setExperimentName<-}
\alias{setExperimentName}
\alias{setCountMatrix<-,scRNAseq-method}
\alias{setCountMatrix<-}
\alias{setCountMatrix}
\alias{setSceNorm<-,scRNAseq-method}
\alias{setSceNorm<-}
\alias{setSceNorm}
\alias{setSpecies<-,scRNAseq-method}
\alias{setSpecies<-}
\alias{setSpecies}
\alias{setOutputDirectory<-,scRNAseq-method}
\alias{setOutputDirectory<-}
\alias{setOutputDirectory}
\alias{setTSNEList<-,scRNAseq-method}
\alias{setTSNEList<-}
\alias{setTSNEList}
\alias{setDbscanList<-,scRNAseq-method}
\alias{setDbscanList<-}
\alias{setDbscanList}
\alias{setCellsSimilarityMatrix<-,scRNAseq-method}
\alias{setCellsSimilarityMatrix<-}
\alias{setCellsSimilarityMatrix}
\alias{setClustersSimilarityMatrix<-,scRNAseq-method}
\alias{setClustersSimilarityMatrix<-}
\alias{setClustersSimilarityMatrix}
\alias{setClustersSimiliratyOrdered<-,scRNAseq-method}
\alias{setClustersSimiliratyOrdered<-}
\alias{setClustersSimiliratyOrdered}
\alias{setMarkerGenesList<-,scRNAseq-method}
\alias{setMarkerGenesList<-}
\alias{setMarkerGenesList}
\alias{setClustersMarkers<-,scRNAseq-method}
\alias{setClustersMarkers<-}
\alias{setClustersMarkers}
\alias{setGenesInfos<-,scRNAseq-method}
\alias{setGenesInfos<-}
\alias{setGenesInfos}
\alias{setName<-,Tsne-method}
\alias{setPC<-,Tsne-method}
\alias{setPC<-}
\alias{setPC}
\alias{setPerplexity<-,Tsne-method}
\alias{setPerplexity<-}
\alias{setPerplexity}
\alias{setCoordinates<-,Tsne-method}
\alias{setCoordinates<-}
\alias{setCoordinates}
\alias{setName<-,Dbscan-method}
\alias{setName<-}
\alias{setEpsilon<-,Dbscan-method}
\alias{setEpsilon<-}
\alias{setEpsilon}
\alias{setMinPoints<-,Dbscan-method}
\alias{setMinPoints<-}
\alias{setMinPoints}
\alias{setClustering<-,Dbscan-method}
\alias{setClustering<-}
\alias{setClustering}
\title{setters}
\usage{
setExperimentName(theObject) <- value

setCountMatrix(theObject) <- value

setSceNorm(theObject) <- value

setSpecies(theObject) <- value

setOutputDirectory(theObject) <- value

setTSNEList(theObject) <- value

setDbscanList(theObject) <- value

setCellsSimilarityMatrix(theObject) <- value

setClustersSimilarityMatrix(theObject) <- value

setClustersSimiliratyOrdered(theObject) <- value

setMarkerGenesList(theObject) <- value

setClustersMarkers(theObject) <- value

setGenesInfos(theObject) <- value

\S4method{setName}{Tsne}(theObject) <- value

setPC(theObject) <- value

setPerplexity(theObject) <- value

setCoordinates(theObject) <- value

setName(theObject) <- value

setEpsilon(theObject) <- value

setMinPoints(theObject) <- value

setClustering(theObject) <- value
}
\arguments{
\item{theObject}{A scRNA-seq, Tsne or Dbscan object to update. See
description or ?scRNAseq, ?Tsne or ?Dbscan.}

\item{value}{The value to update the slot with. See ?scRNAseq, ?Tsne or
?Dbscan.}
}
\value{
setExperimentName: Update the experiment name slot with a character
string (scRNA-seq).

setCountMatrix: Update the countMatrix slot with a matrix of numeric
(scRNA-seq).

setSceNorm: Update the normalized countMatrix slot with SingleCellExperiment
object (scRNA-seq).

setSpecies: Update the species slot with a character string. Value should be
mouse or human. Other organisms can be added on demand  (scRNA-seq).

setOutputDirectory: Update the outputDirectory slot with a character string.
Value should be a valid path (scRNA-seq).

setTSNEList: Update the tSNEList slot with a list of tSNE objects. See
?Tsne-class (scRNA-seq).

setDbscanList: Update the dbscanList slot with a list of dbscan objects. See
?Dbscan-class (scRNA-seq).

setCellsSimilarityMatrix: Update the cellsSimilarityMatrix slot with a
numeric matrix (scRNA-seq).

setClustersSimilarityMatrix: Update the clustersSimilarityMatrix slot with a
numeric matrix (scRNA-seq).

setClustersSimiliratyOrdered: Update the clustersSimilarityOrdered slot with
a numeric factor (scRNA-seq).

setMarkerGenesList: Update the markerGenesList slot with a list of data
frame. The data frame structure should be:
data.frame(Gene = c("gene1"), mean_log10_fdr = c(NA), n_05 = c(NA),
score = c(NA))  (scRNA-seq).

setClustersMarkers: Update the clustersMarkers slot with a data frame.
The data frame structure should be:
data.frame(geneName="gene1", clusters=NA). (scRNA-seq)

setGenesInfos: Update the genesInfos slot with a data frame.
The data frame structure should be:
data.frame(uniprot_gn_symbol=c("symbol"), clusters="1",
                external_gene_name="gene", go_id="GO1,GO2",
                mgi_description="description",
                entrezgene_description="descr",
                gene_biotype="gene", chromosome_name="1", Symbol="symbol",
                ensembl_gene_id="ENS", mgi_id="MGI", entrezgene_id="1",
                uniprot_gn_id="ID"). (scRNA-seq)

setPC: Update the pc slot with a vector of numeric (Tsne).

setPerplexity: Update the perplexity slot with a vector of numeric (Tsne).

setCoordinates: Update the coordinates slot with a matrix of numeric (Tsne).

setName: Update the Tsne or Dbscan name slot with a character string
(Dbscan).

setEpsilon: Update the epsilon slot with a vector of numeric (Dbscan).

setMinPoints: Update the minPoints slot with a vector of numeric (Dbscan).

setClustering: Update the clustering slot with a matrix of numeric (Dbscan).
}
\description{
Update a slot of a scRNA-seq, Tsne or Dbscan object.
}
\examples{
## Load the count matrix
countmatrixPath <- system.file("extdata/example_countMatrix.tsv",
                            package="conclus")
countMatrix <- loadDataOrMatrix(file=countmatrixPath, type="countMatrix")

## Load the coldata
coldataPath <- system.file("extdata/example_colData.tsv",
                            package="conclus")
columnsMetaData <- loadDataOrMatrix(file=coldataPath, type="coldata",
columnID="cell_ID")

## Create the initial object
scr <- singlecellRNAseq(experimentName = "Bergiers",
                countMatrix     = countMatrix,
                species         = "mouse",
                outputDirectory = "YourOutputDirectory")

setExperimentName(scr) <- "newName"
setCountMatrix(scr) <- countMatrix[seq_len(15), seq_len(100)]
setSpecies(scr) <- "human"
setOutputDirectory(scr) <- "newPath"

}
